Module rgsl::airy

Functions

Ai

This routine computes the Airy function Ai(x) with an accuracy specified by mode.

Ai_deriv

This routine computes the Airy function derivative Ai’(x) with an accuracy specified by mode.

Ai_deriv_e

This routine computes the Airy function derivative Ai’(x) with an accuracy specified by mode.

Ai_deriv_scaled

This routine computes the scaled Airy function derivative S_A(x) Ai’(x). For x>0 the scaling factor S_A(x) is \exp(+(2/3) x^(3/2)), and is 1 for x<0.

Ai_deriv_scaled_e

This routine computes the scaled Airy function derivative S_A(x) Ai’(x). For x>0 the scaling factor S_A(x) is \exp(+(2/3) x^(3/2)), and is 1 for x<0.

Ai_e

This routine computes the Airy function Ai(x) with an accuracy specified by mode.

Ai_scaled

This routine computes a scaled version of the Airy function S_A(x) Ai(x). For x>0 the scaling factor S_A(x) is \exp(+(2/3) x^(3/2)), and is 1 for x<0.

Ai_scaled_e

This routine computes a scaled version of the Airy function S_A(x) Ai(x). For x>0 the scaling factor S_A(x) is \exp(+(2/3) x^(3/2)), and is 1 for x<0.

Bi

This routine computes the Airy function Bi(x) with an accuracy specified by mode.

Bi_deriv

This routine computes the Airy function derivative Bi’(x) with an accuracy specified by mode.

Bi_deriv_e

This routine computes the Airy function derivative Bi’(x) with an accuracy specified by mode.

Bi_deriv_scaled

This routine computes the scaled Airy function derivative S_B(x) Bi’(x). For x>0 the scaling factor S_B(x) is exp(-(2/3) x^(3/2)), and is 1 for x<0.

Bi_deriv_scaled_e

This routine computes the scaled Airy function derivative S_B(x) Bi’(x). For x>0 the scaling factor S_B(x) is exp(-(2/3) x^(3/2)), and is 1 for x<0.

Bi_e

This routine computes the Airy function Bi(x) with an accuracy specified by mode.

Bi_scaled

This routine computes a scaled version of the Airy function S_B(x) Bi(x). For x>0 the scaling factor S_B(x) is exp(-(2/3) x^(3/2)), and is 1 for x<0.

Bi_scaled_e

This routine computes a scaled version of the Airy function S_B(x) Bi(x). For x>0 the scaling factor S_B(x) is exp(-(2/3) x^(3/2)), and is 1 for x<0.

zero_Ai

This routine computes the location of the s-th zero of the Airy function Ai(x).

zero_Ai_deriv

This routine computes the location of the s-th zero of the Airy function derivative Ai’(x).

zero_Ai_deriv_e

This routine computes the location of the s-th zero of the Airy function derivative Ai’(x).

zero_Ai_e

This routine computes the location of the s-th zero of the Airy function Ai(x).

zero_Bi

This routine computes the location of the s-th zero of the Airy function Bi(x).

zero_Bi_deriv

This routine computes the location of the s-th zero of the Airy function derivative Bi’(x).

zero_Bi_deriv_e

This routine computes the location of the s-th zero of the Airy function derivative Bi’(x).

zero_Bi_e

This routine computes the location of the s-th zero of the Airy function Bi(x).