Function rgsl::blas::level2::cher2

pub fn cher2(uplo: Uplo, alpha: &ComplexF32, x: &VectorComplexF32, y: &VectorComplexF32, A: &mut MatrixComplexF32) -> Value

These functions compute the hermitian rank-2 update A = \alpha x yH + \alpha* y xH + A of the hermitian matrix A. Since the matrix A is hermitian only its upper half or lower half need to be stored. When Uplo is CblasUpper then the upper triangle and diagonal of A are used, and when Uplo is CblasLower then the lower triangle and diagonal of A are used. The imaginary elements of the diagonal are automatically set to zero.