Function rgsl::blas::level3::csyrk

pub fn csyrk(uplo: Uplo, trans: Transpose, alpha: &ComplexF32, A: &MatrixComplexF32, beta: &ComplexF32, C: &mut MatrixComplexF32) -> Value

This function computes a rank-k update of the symmetric matrix C, C = \alpha A AT + \beta C when Trans is NoTrans and C = \alpha AT A + \beta C when Trans is Trans. Since the matrix C is symmetric only its upper half or lower half need to be stored. When Uplo is Upper then the upper triangle and diagonal of C are used, and when Uplo is Lower then the lower triangle and diagonal of C are used.