Module rgsl::coupling_coefficients

The Wigner 3-j, 6-j and 9-j symbols give the coupling coefficients for combined angular momentum vectors. Since the arguments of the standard coupling coefficient functions are integer or half-integer, the arguments of the following functions are, by convention, integers equal to twice the actual spin value.

Functions

_3j
This routine computes the Wigner 3-j coefficient,

_3j_e
This routine computes the Wigner 3-j coefficient,

_6j
This routine computes the Wigner 6-j coefficient,

_6j_e
This routine computes the Wigner 6-j coefficient,

_9j
This routine computes the Wigner 9-j coefficient,

_9j_e
This routine computes the Wigner 9-j coefficient,