Function rgsl::eigen::nonsymmv_sort

pub fn nonsymmv_sort(eval: &VectorComplexF64, evec: &MatrixComplexF64, sort_type: EigenSort) -> Value

This function simultaneously sorts the eigenvalues stored in the vector eval and the corresponding complex eigenvectors stored in the columns of the matrix evec into ascending or descending order according to the value of the parameter sort_type. Only EigenSort::AbsAsc and EigenSort::AbsDesc are supported due to the eigenvalues being complex.