Module rgsl::elliptic::legendre

The Legendre forms of elliptic integrals F(\phi,k), E(\phi,k) and \Pi(\phi,k,n) are defined by,

F(\phi,k) = \int_0\phi dt 1/\sqrt((1 - k2 \sin2(t)))

E(\phi,k) = \int_0\phi dt \sqrt((1 - k2 \sin2(t)))

Pi(\phi,k,n) = \int_0\phi dt 1/((1 + n \sin2(t))\sqrt(1 - k2 \sin2(t)))

The complete Legendre forms are denoted by K(k) = F(\pi/2, k) and E(k) = E(\pi/2, k).

The notation used here is based on Carlson, Numerische Mathematik 33 (1979) 1 and differs slightly from that used by Abramowitz & Stegun, where the functions are given in terms of the parameter m = k2 and n is replaced by -n.

Modules

complete

incomplete