Function rgsl::elliptic::legendre::complete::ellint_Kcomp

pub fn ellint_Kcomp(k: f64, mode: Mode) -> f64

This routine computes the complete elliptic integral K(k) to the accuracy specified by the mode variable mode. Note that Abramowitz & Stegun define this function in terms of the parameter m = k^2.