Module rgsl::exponential

Functions

exp
This routine provides an exponential function \exp(x) using GSL semantics and error checking.

exp_e
This routine provides an exponential function \exp(x) using GSL semantics and error checking.

exp_e10_e
This function computes the exponential \exp(x) using the gsl_sf_result_e10 type to return a result with extended range. This function may be useful if the value of \exp(x) would overflow the numeric range of double.

exp_err_e
This function exponentiates x with an associated absolute error dx.

exp_err_e10_e
This function exponentiates a quantity x with an associated absolute error dx using the ::types::ResultE10 type to return a result with extended range.

exp_mult
This routine exponentiates x and multiply by the factor y to return the product y \exp(x).

exp_mult_e
This routine exponentiates x and multiply by the factor y to return the product y \exp(x).

exp_mult_e10_e
This function computes the exponential \exp(x) using the gsl_sf_result_e10 type to return a result with extended range. This function may be useful if the value of \exp(x) would overflow the numeric range of double.

exp_mult_err_e
This routine computes the product y \exp(x) for the quantities x, y with associated absolute errors dx, dy.

exp_mult_err_e10_e
This routine computes the product y \exp(x) for the quantities x, y with associated absolute errors dx, dy using the gsl_sf_result_e10 type to return a result with extended range.

expm1
This routine computes the quantity \exp(x)-1 using an algorithm that is accurate for small x.

expm1_e
This routine computes the quantity \exp(x)-1 using an algorithm that is accurate for small x.

exprel
This routine computes the quantity (\exp(x)-1)/x using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion (\exp(x)-1)/x = 1 + x/2 + x2/(2*3) + x3/(2*3*4) + \dots.

exprel_2
This routine computes the quantity 2(\exp(x)-1-x)/x2 using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion 2(\exp(x)-1-x)/x2 = 1 + x/3 + x2/(3*4) + x3/(3*4*5) + \dots.

exprel_2_e
This routine computes the quantity 2(\exp(x)-1-x)/x2 using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion 2(\exp(x)-1-x)/x2 = 1 + x/3 + x2/(3*4) + x3/(3*4*5) + \dots.

exprel_e
This routine computes the quantity (\exp(x)-1)/x using an algorithm that is accurate for small x. For small x the algorithm is based on the expansion (\exp(x)-1)/x = 1 + x/2 + x2/(2*3) + x3/(2*3*4) + \dots.

exprel_n
This routine computes the N-relative exponential, which is the n-th generalization of the functions gsl_sf_exprel and gsl_sf_exprel_2. The N-relative exponential is given by:

exprel_n_e
This routine computes the N-relative exponential, which is the n-th generalization of the functions gsl_sf_exprel and gsl_sf_exprel_2. The N-relative exponential is given by: