Module rgsl::exponential_integrals

Functions

Chi
This routine computes the integral Chi(x) := \Re[ \gamma_E + \log(x) + \int_0x dt (\cosh(t)-1)/t] , where \gamma_E is the Euler constant (available as the macro M_EULER).

Chi_e
This routine computes the integral Chi(x) := \Re[ \gamma_E + \log(x) + \int_0x dt (\cosh(t)-1)/t] , where \gamma_E is the Euler constant (available as the macro M_EULER).

Ci
This routine computes the Cosine integral Ci(x) = -\int_x\infty dt \cos(t)/t for x > 0.

Ci_e
This routine computes the Cosine integral Ci(x) = -\int_x\infty dt \cos(t)/t for x > 0.

E1
This routine computes the exponential integral E_1(x),

E1_e
This routine computes the exponential integral E_1(x),

E2
This routine computes the second-order exponential integral E_2(x),

E2_e
This routine computes the second-order exponential integral E_2(x),

Ei
This routine computes the exponential integral Ei(x),

Ei_e
This routine computes the exponential integral Ei(x),

En
This routine computes the exponential integral E_n(x) of order n,

En_e
This routine computes the exponential integral E_n(x) of order n,

Shi
This routine computes the integral Shi(x) = \int_0x dt \sinh(t)/t.

Shi_e
This routine computes the integral Shi(x) = \int_0x dt \sinh(t)/t.

Si
This routine computes the Sine integral Si(x) = \int_0x dt \sin(t)/t.

Si_e
This routine computes the Sine integral Si(x) = \int_0x dt \sin(t)/t.

_3
This routine computes the third-order exponential integral Ei_3(x) = \int_0xdt \exp(-t3) for x >= 0.

_3_e
This routine computes the third-order exponential integral Ei_3(x) = \int_0xdt \exp(-t3) for x >= 0.

atanint
This routine computes the Arctangent integral, which is defined as AtanInt(x) = \int_0x dt \arctan(t)/t.

atanint_e
This routine computes the Arctangent integral, which is defined as AtanInt(x) = \int_0x dt \arctan(t)/t.