Module rgsl::fft::mixed_radix

These functions compute forward, backward and inverse FFTs of length n with stride stride, on the packed complex array data, using a mixed radix decimation-in-frequency algorithm. There is no restriction on the length n. Efficient modules are provided for subtransforms of length 2, 3, 4, 5, 6 and 7. Any remaining factors are computed with a slow, O(n2), general-n module. The caller must supply a wavetable containing the trigonometric lookup tables and a workspace work. For the transform version of the function the sign argument can be either forward (-1) or backward (+1).

The functions return a value of 0 if no errors were detected. The following gsl_errno conditions are defined for these functions:

Value::Dom

The length of the data n is not a positive integer (i.e. n is zero).

Value::Invalid

The length of the data n and the length used to compute the given wavetable do not match.

Functions

backward

forward

inverse

transform