Module rgsl::fft::radix2

These functions compute forward, backward and inverse FFTs of length n with stride stride, on the packed complex array data using an in-place radix-2 decimation-in-time algorithm. The length of the transform is restricted to powers of two. For the transform version of the function the sign argument can be either forward (-1) or backward (+1).

The functions return a value of ::Value::Success if no errors were detected, or Value::Dom if the length n is not a power of two.

Functions

backward

dif_backward
This is decimation-in-frequency version of the radix-2 FFT function.

dif_forward
This is decimation-in-frequency version of the radix-2 FFT function.

dif_inverse
This is decimation-in-frequency version of the radix-2 FFT function.

dif_transform
This is decimation-in-frequency version of the radix-2 FFT function.

forward

inverse

transform