Module rgsl::gamma_beta::beta

Functions

beta
This routine computes the Beta Function, B(a,b) = \Gamma(a)\Gamma(b)/\Gamma(a+b) subject to a and b not being negative integers.

beta_e
This routine computes the Beta Function, B(a,b) = \Gamma(a)\Gamma(b)/\Gamma(a+b) subject to a and b not being negative integers.

lnbeta
This routine computes the logarithm of the Beta Function, \log(B(a,b)) subject to a and b not being negative integers.

lnbeta_e
This routine computes the logarithm of the Beta Function, \log(B(a,b)) subject to a and b not being negative integers.