Function rgsl::interpolation::eval_deriv_e

pub fn eval_deriv_e(
    interp: &Interp, 
    xa: &[f64], 
    ya: &[f64], 
    x: f64, 
    acc: &mut InterpAccel, 
    d: &mut f64
) -> Value

This function returns the derivative d of an interpolated function for a given point x, using the interpolation object interp, data arrays xa and ya and the accelerator acc.