Module rgsl::laguerre

The generalized Laguerre polynomials are defined in terms of confluent hypergeometric functions as La_n(x) = ((a+1)_n / n!) 1F1(-n,a+1,x), and are sometimes referred to as the associated Laguerre polynomials. They are related to the plain Laguerre polynomials L_n(x) by L0_n(x) = L_n(x) and Lk_n(x) = (-1)k (dk/dxk) L_(n+k)(x). For more information see Abramowitz & Stegun, Chapter 22.

Functions

laguerre_1
This function evaluates the generalized Laguerre polynomials La_1(x), La_2(x), La_3(x) using explicit representations.

laguerre_1_e
This function evaluates the generalized Laguerre polynomials La_1(x), La_2(x), La_3(x) using explicit representations.

laguerre_2
This function evaluates the generalized Laguerre polynomials La_1(x), La_2(x), La_3(x) using explicit representations.

laguerre_2_e
This function evaluates the generalized Laguerre polynomials La_1(x), La_2(x), La_3(x) using explicit representations.

laguerre_3
This function evaluates the generalized Laguerre polynomials La_1(x), La_2(x), La_3(x) using explicit representations.

laguerre_3_e
This function evaluates the generalized Laguerre polynomials La_1(x), La_2(x), La_3(x) using explicit representations.

laguerre_n
the generalized Laguerre polynomials La_n(x) for a > -1, n >= 0.

laguerre_n_e
the generalized Laguerre polynomials La_n(x) for a > -1, n >= 0.