Module rgsl::legendre::radial

The following spherical functions are specializations of Legendre functions which give the regular eigenfunctions of the Laplacian on a 3-dimensional hyperbolic space H3d. Of particular interest is the flat limit, \lambda \to \infty, \eta \to 0, \lambda\eta fixed.

Functions

legendre_H3d
This routine computes the l-th radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space \eta >= 0, l >= 0. In the flat limit this takes the form L{H3d}_l(\lambda,\eta) = j_l(\lambda\eta).

legendre_H3d_0
This routine computes the zeroth radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space, L{H3d}_0(\lambda,\eta) := \sin(\lambda\eta)/(\lambda\sinh(\eta)) for \eta >= 0. In the flat limit this takes the form L{H3d}_0(\lambda,\eta) = j_0(\lambda\eta).

legendre_H3d_0_e
This routine computes the zeroth radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space, L{H3d}_0(\lambda,\eta) := \sin(\lambda\eta)/(\lambda\sinh(\eta)) for \eta >= 0. In the flat limit this takes the form L{H3d}_0(\lambda,\eta) = j_0(\lambda\eta).

legendre_H3d_1
This routine computes the first radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space, L{H3d}_1(\lambda,\eta) := 1/\sqrt{\lambda2 + 1} \sin(\lambda \eta)/(\lambda \sinh(\eta)) (\coth(\eta) - \lambda \cot(\lambda\eta)) for \eta >= 0. In the flat limit this takes the form L{H3d}_1(\lambda,\eta) = j_1(\lambda\eta).

legendre_H3d_1_e
This routine computes the first radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space, L{H3d}_1(\lambda,\eta) := 1/\sqrt{\lambda2 + 1} \sin(\lambda \eta)/(\lambda \sinh(\eta)) (\coth(\eta) - \lambda \cot(\lambda\eta)) for \eta >= 0. In the flat limit this takes the form L{H3d}_1(\lambda,\eta) = j_1(\lambda\eta).

legendre_H3d_array
This function computes an array of radial eigenfunctions L{H3d}_l(\lambda, \eta) for 0 <= l <= lmax.

legendre_H3d_e
This routine computes the l-th radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space \eta >= 0, l >= 0. In the flat limit this takes the form L{H3d}_l(\lambda,\eta) = j_l(\lambda\eta).