Function rgsl::legendre::radial::legendre_H3d

pub fn legendre_H3d(l: i32, lambda: f64, eta: f64) -> f64

This routine computes the l-th radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space \eta >= 0, l >= 0. In the flat limit this takes the form L{H3d}_l(\lambda,\eta) = j_l(\lambda\eta).