Function rgsl::legendre::radial::legendre_H3d_0_e

pub fn legendre_H3d_0_e(lambda: f64, eta: f64) -> (Value, Result)

This routine computes the zeroth radial eigenfunction of the Laplacian on the 3-dimensional hyperbolic space, L^{H3d}_0(\lambda,\eta) := \sin(\lambda\eta)/(\lambda\sinh(\eta)) for \eta >= 0. In the flat limit this takes the form L^{H3d}_0(\lambda,\eta) = j_0(\lambda\eta).