Function rgsl::linear_algebra::cholesky_decomp

pub fn cholesky_decomp(a: &MatrixF64) -> Value

This function factorizes the symmetric, positive-definite square matrix A into the Cholesky decomposition A = L LT (or A = L LH for the complex case). On input, the values from the diagonal and lower-triangular part of the matrix A are used (the upper triangular part is ignored). On output the diagonal and lower triangular part of the input matrix A contain the matrix L, while the upper triangular part of the input matrix is overwritten with LT (the diagonal terms being identical for both L and LT). If the matrix is not positive-definite then the decomposition will fail, returning the error code ::Dom.

When testing whether a matrix is positive-definite, disable the error handler first to avoid triggering an error.