Function rgsl::linear_algebra::cholesky_svx

pub fn cholesky_svx(cholesky: &MatrixF64, x: &VectorF64) -> Value

This function solves the system A x = b in-place using the Cholesky decomposition of A held in the matrix cholesky which must have been previously computed by gsl_linalg_cholesky_decomp or gsl_linalg_complex_cholesky_decomp. On input x should contain the right-hand side b, which is replaced by the solution on output.