Function rgsl::linear_algebra::complex_LU_decomp

pub fn complex_LU_decomp(
    a: &mut MatrixComplexF64, 
    p: &mut Permutation, 
    signum: &mut i32
) -> Value

Factorise a general N x N complex matrix A into,

P A = L U

where P is a permutation matrix, L is unit lower triangular and U is upper triangular.

L is stored in the strict lower triangular part of the input matrix. The diagonal elements of L are unity and are not stored.

U is stored in the diagonal and upper triangular part of the input matrix.

P is stored in the permutation p. Column j of P is column k of the identity matrix, where k = permutation->data[j]

signum gives the sign of the permutation, (-1)^n, where n is the number of interchanges in the permutation.

See Golub & Van Loan, Matrix Computations, Algorithm 3.4.1 (Gauss Elimination with Partial Pivoting).