Function rgsl::linear_algebra::complex_householder_hv

pub fn complex_householder_hv(
    tau: &ComplexF64, 
    v: &VectorComplexF64, 
    w: &mut MatrixComplexF64
) -> Value

This function applies the Householder transformation P defined by the scalar tau and the vector v to the vector w. On output the result P w is stored in w.