Function rgsl::linear_algebra::complex_householder_transform

pub fn complex_householder_transform(v: &VectorComplexF64) -> ComplexF64

This function prepares a Householder transformation P = I - \tau v vT which can be used to zero all the elements of the input vector except the first. On output the transformation is stored in the vector v and the scalar \tau is returned.