Function rgsl::linear_algebra::hermtd_unpack

pub fn hermtd_unpack(
    a: &MatrixComplexF64, 
    tau: &VectorComplexF64, 
    u: &mut MatrixComplexF64, 
    diag: &mut VectorF64, 
    subdiag: &mut VectorF64
) -> Value

This function unpacks the encoded tridiagonal decomposition (A, tau) obtained from gsl_linalg_hermtd_decomp into the unitary matrix U, the real vector of diagonal elements diag and the real vector of subdiagonal elements subdiag.