Function rgsl::linear_algebra::hessenberg_unpack

pub fn hessenberg_unpack(h: &MatrixF64, tau: &VectorF64, u: &MatrixF64) -> Value

This function constructs the orthogonal matrix U from the information stored in the Hessenberg matrix H along with the vector tau. H and tau are outputs from gsl_linalg_hessenberg_decomp.