Module rgsl::logarithm

Information on the properties of the Logarithm function can be found in Abramowitz & Stegun, Chapter 4.

Functions

complex_log_e
This routine computes the complex logarithm of z = z_r + i z_i. The results are returned as lnr, theta such that \exp(lnr + i \theta) = z_r + i z_i, where \theta lies in the range [-\pi,\pi].

log
This routine computes the logarithm of x, \log(x), for x > 0.

log_1plusx
This routine computes \log(1 + x) for x > -1 using an algorithm that is accurate for small x.

log_1plusx_e
This routine computes \log(1 + x) for x > -1 using an algorithm that is accurate for small x.

log_1plusx_mx
This routine computes \log(1 + x) - x for x > -1 using an algorithm that is accurate for small x.

log_1plusx_mx_e
This routine computes \log(1 + x) - x for x > -1 using an algorithm that is accurate for small x.

log_abs
This routine computes the logarithm of the magnitude of x, \log(|x|), for x \ne 0.

log_abs_e
This routine computes the logarithm of the magnitude of x, \log(|x|), for x \ne 0.

log_e
This routine computes the logarithm of x, \log(x), for x > 0.