Function rgsl::logarithm::complex_log_e

pub fn complex_log_e(zr: f64, zi: f64) -> (Value, Result, Result)

This routine computes the complex logarithm of z = z_r + i z_i. The results are returned as lnr, theta such that \exp(lnr + i \theta) = z_r + i z_i, where \theta lies in the range [-\pi,\pi].