Function rgsl::polynomials::cubic_equations::poly_solve_cubic

pub fn poly_solve_cubic(a: f64, b: f64, c: f64, x0: &mut f64, x1: &mut f64, x2: &mut f64) -> i32

This function finds the real roots of the cubic equation,

x3 + a x2 + b x + c = 0

with a leading coefficient of unity. The number of real roots (either one or three) is returned, and their locations are stored in x0, x1 and x2. If one real root is found then only x0 is modified. When three real roots are found they are stored in x0, x1 and x2 in ascending order. The case of coincident roots is not considered special. For example, the equation (x-1)3=0 will have three roots with exactly equal values. As in the quadratic case, finite precision may cause equal or closely-spaced real roots to move off the real axis into the complex plane, leading to a discrete change in the number of real roots.