Function rgsl::polynomials::divided_difference_representation::poly_dd_init

pub fn poly_dd_init(dd: &mut [f64], xa: &[f64], ya: &[f64]) -> Value

This function computes a divided-difference representation of the interpolating polynomial for the points (x, y) stored in the arrays xa and ya of length size. On output the divided-differences of (xa,ya) are stored in the array dd, also of length size. Using the notation above, dd[k] = [x_0,x_1,...,x_k].