Function rgsl::polynomials::quadratic_equations::poly_complex_solve_quadratic

pub fn poly_complex_solve_quadratic(
    a: f64, 
    b: f64, 
    c: f64, 
    z0: &mut ComplexF64, 
    z1: &mut ComplexF64
) -> i32

This function finds the complex roots of the quadratic equation,

a z^2 + b z + c = 0

The number of complex roots is returned (either one or two) and the locations of the roots are stored in z0 and z1. The roots are returned in ascending order, sorted first by their real components and then by their imaginary components. If only one real root is found (i.e. if a=0) then it is stored in z0.