Module rgsl::power

The following functions are equivalent to the function gsl_pow_int (see 1Small integer powers1) with an error estimate.

Functions

pow_int
This routine computes the power xn for integer n. The power is computed using the minimum number of multiplications. For example, x8 is computed as ((x2)2)2, requiring only 3 multiplications. For reasons of efficiency, these functions do not check for overflow or underflow conditions.

pow_int_e
This routine computes the power xn for integer n. The power is computed using the minimum number of multiplications. For example, x8 is computed as ((x2)2)2, requiring only 3 multiplications. For reasons of efficiency, these functions do not check for overflow or underflow conditions.