Module rgsl::randist::exponential

Functions

exponential
This function returns a random variate from the exponential distribution with mean mu. The distribution is,

exponential_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the exponential distribution with mean mu.

exponential_Pinv

exponential_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the exponential distribution with mean mu.

exponential_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the exponential distribution with mean mu.

exponential_pdf
This function computes the probability density p(x) at x for an exponential distribution with mean mu, using the formula given above.