Module rgsl::randist::gaussian

Functions

gaussian
This function returns a Gaussian random variate, with mean zero and standard deviation sigma. The probability distribution for Gaussian random variates is,

gaussian_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Gaussian distribution with standard deviation sigma.

gaussian_pdf
This function computes the probability density p(x) at x for a Gaussian distribution with standard deviation sigma, using the formula given above.

gaussian_ratio_method
This function computes a Gaussian random variate using the alternative Marsaglia-Tsang ziggurat and Kinderman-Monahan-Leva ratio methods. The Ziggurat algorithm is the fastest available algorithm in most cases.

gaussian_ziggurat

ugaussian
This function computes results for the unit Gaussian distribution. They are equivalent to the functions above with a standard deviation of one, sigma = 1.

ugaussian_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the unit Gaussian distribution.

ugaussian_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the unit Gaussian distribution.

ugaussian_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the unit Gaussian distribution.

ugaussian_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the unit Gaussian distribution.

ugaussian_pdf
This function computes results for the unit Gaussian distribution. They are equivalent to the functions above with a standard deviation of one, sigma = 1.

ugaussian_ratio_method
This function computes results for the unit Gaussian distribution. They are equivalent to the functions above with a standard deviation of one, sigma = 1.