Module rgsl::randist::geometric

Functions

geometric
This function returns a random integer from the geometric distribution, the number of independent trials with probability p until the first success. The probability distribution for geometric variates is,

geometric_P
This function computes the cumulative distribution functions P(k), Q(k) for the geometric distribution with parameter p.

geometric_Q
This function computes the cumulative distribution functions P(k), Q(k) for the geometric distribution with parameter p.

geometric_pdf
This function computes the probability p(k) of obtaining k from a geometric distribution with probability parameter p, using the formula given above.