Module rgsl::randist::hypergeometric

Functions

hypergeometric
This function returns a random integer from the hypergeometric distribution. The probability distribution for hypergeometric random variates is,

hypergeometric_P
This function computes the cumulative distribution functions P(k), Q(k) for the hypergeometric distribution with parameters n1, n2 and t.

hypergeometric_Q
This function computes the cumulative distribution functions P(k), Q(k) for the hypergeometric distribution with parameters n1, n2 and t.

hypergeometric_pdf
This function computes the probability p(k) of obtaining k from a hypergeometric distribution with parameters n1, n2, t, using the formula given above.