Module rgsl::randist::lognormal

Functions

lognormal
This function returns a random variate from the lognormal distribution. The distribution function is,

lognormal_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the lognormal distribution with parameters zeta and sigma.

lognormal_pdf
This function computes the probability density p(x) at x for a lognormal distribution with parameters zeta and sigma, using the formula given above.