Module rgsl::randist::pareto

Functions

pareto
This function returns a random variate from the Pareto distribution of order a. The distribution function is,

pareto_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Pareto distribution with exponent a and scale b.

pareto_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Pareto distribution with exponent a and scale b.

pareto_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Pareto distribution with exponent a and scale b.

pareto_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Pareto distribution with exponent a and scale b.

pareto_pdf
This function computes the probability density p(x) at x for a Pareto distribution with exponent a and scale b, using the formula given above.