Module rgsl::randist::rayleigh

Functions

rayleigh
This function returns a random variate from the Rayleigh distribution with scale parameter sigma. The distribution is,

rayleigh_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Rayleigh distribution with scale parameter sigma.

rayleigh_pdf
This function computes the probability density p(x) at x for a Rayleigh distribution with scale parameter sigma, using the formula given above.