Function rgsl::randist::rayleigh_tail::rayleigh_tail

pub fn rayleigh_tail(r: &Rng, a: f64, sigma: f64) -> f64

This function returns a random variate from the tail of the Rayleigh distribution with scale parameter sigma and a lower limit of a. The distribution is,

p(x) dx = {x \over \sigma2} \exp ((a2 - x2) /(2 \sigma2)) dx

for x > a.