Function rgsl::randist::shuffling_sampling::sample

pub fn sample<T>(r: &Rng, dest: &mut [T], src: &[T]) -> Value

This function is like gsl_ran_choose but samples k items from the original array of n items src with replacement, so the same object can appear more than once in the output sequence dest. There is no requirement that k be less than n in this case.