Function rgsl::randist::spherical_vector::dir_2d_trig_method

pub fn dir_2d_trig_method(r: &Rng, x: &mut f64, y: &mut f64)

This function returns a random direction vector v = (x,y) in two dimensions. The vector is normalized such that |v|2 = x2 + y2 = 1. The obvious way to do this is to take a uniform random number between 0 and 2\pi and let x and y be the sine and cosine respectively. Two trig functions would have been expensive in the old days, but with modern hardware implementations, this is sometimes the fastest way to go. This is the case for the Pentium (but not the case for the Sun Sparcstation). One can avoid the trig evaluations by choosing x and y in the interior of a unit circle (choose them at random from the interior of the enclosing square, and then reject those that are outside the unit circle), and then dividing by \sqrt{x2 + y2}. A much cleverer approach, attributed to von Neumann (See Knuth, v2, 3rd ed, p140, exercise 23), requires neither trig nor a square root. In this approach, u and v are chosen at random from the interior of a unit circle, and then x=(u2-v2)/(u2+v2) and y=2uv/(u2+v2).