Function rgsl::sort::vectors::sort_index

pub fn sort_index(p: &mut [usize], data: &[f64], stride: usize, n: usize)

This function indirectly sorts the n elements of the array data with stride stride into ascending order, storing the resulting permutation in p. The array p must be allocated with a sufficient length to store the n elements of the permutation. The elements of p give the index of the array element which would have been stored in that position if the array had been sorted in place. The array data is not changed.