Function rgsl::sort::vectors::sort_vector_index

pub fn sort_vector_index(p: &Permutation, v: &VectorF64) -> Value

This function indirectly sorts the elements of the vector v into ascending order, storing the resulting permutation in p. The elements of p give the index of the vector element which would have been stored in that position if the vector had been sorted in place. The first element of p gives the index of the least element in v, and the last element of p gives the index of the greatest element in v. The vector v is not changed.