Function rgsl::statistics::variance_with_fixed_mean

pub fn variance_with_fixed_mean(data: &[f64], stride: usize, n: usize, mean: f64) -> f64

This function computes an unbiased estimate of the variance of data when the population mean mean of the underlying distribution is known a priori. In this case the estimator for the variance uses the factor 1/N and the sample mean \Hat\mu is replaced by the known population mean \mu,

\Hat\sigma2 = (1/N) \sum (x_i - \mu)2