Module rgsl::types

Reexports

pub use self::basis_spline::{BSpLineWorkspace, BSpLineDerivWorkspace};

pub use self::chebyshev::ChebSeries;

pub use self::combination::Combination;

pub use self::complex::{ComplexF32, ComplexF64};

pub use self::discrete_hankel::DiscreteHankel;

pub use self::eigen_symmetric_workspace::{EigenSymmetricWorkspace, EigenSymmetricVWorkspace, EigenHermitianWorkspace, EigenHermitianVWorkspace, EigenNonSymmWorkspace, EigenNonSymmVWorkspace, EigenGenSymmWorkspace, EigenGenSymmVWorkspace, EigenGenHermWorkspace, EigenGenHermVWorkspace, EigenGenWorkspace, EigenGenVWorkspace};

pub use self::fast_fourier_transforms::{FftComplexWaveTable, FftComplexWorkspace};

pub use self::histograms::{Histogram, HistogramPdf, Histogram2D, Histogram2DPdf};

pub use self::integration::{IntegrationWorkspace, IntegrationQawsTable, IntegrationQawoTable, CquadWorkspace, GLFixedTable};

pub use self::interpolation::{InterpAccel, Interp, InterpType, Spline};

pub use self::mathieu::MathieuWorkspace;

pub use self::matrix::{MatrixF32, MatrixF64, MatrixView};

pub use self::matrix_complex::{MatrixComplexF32, MatrixComplexF64};

pub use self::minimizer::{Minimizer, MinimizerType};

pub use self::monte_carlo::{PlainMonteCarlo, MiserMonteCarlo, MiserParams, VegasMonteCarlo};

pub use self::multifit_solver::{MultiFitFdfSolver, MultiFitFunction, MultiFitFdfSolverType, MultiFitFunctionFdf};

pub use self::multiset::MultiSet;

pub use self::n_tuples::NTuples;

pub use self::ordinary_differential_equations::{ODEiv2System, ODEiv2Step, ODEiv2StepType, ODEiv2Control, ODEiv2Evolve, ODEiv2Driver};

pub use self::permutation::Permutation;

pub use self::polynomial::PolyComplex;

pub use self::qrng::{QRng, QRngType};

pub use self::ran_discrete::RanDiscrete;

pub use self::result::{Result, ResultE10};

pub use self::rng::{Rng, RngType};

pub use self::series_acceleration::{LevinUWorkspace, LevinUTruncWorkspace};

pub use self::vector::{VectorF32, VectorF64, VectorView};

pub use self::vector_complex::{VectorComplexF32, VectorComplexF64};

pub use self::wavelet_transforms::{Wavelet, WaveletType, WaveletWorkspace};

Modules

basis_spline
B-splines are commonly used as basis functions to fit smoothing curves to large data sets. To do this, the abscissa axis is broken up into some number of intervals, where the endpoints of each interval are called breakpoints. These breakpoints are then converted to knots by imposing various continuity and smoothness conditions at each interface. Given a nondecreasing knot vector t = {t_0, t_1, …, t_{n+k-1}}, the n basis splines of order k are defined by

chebyshev

combination

complex

discrete_hankel

eigen_symmetric_workspace

fast_fourier_transforms

histograms

integration

interpolation

mathieu
The routines described in this section compute the angular and radial Mathieu functions, and their characteristic values. Mathieu functions are the solutions of the following two differential equations:

matrix

matrix_complex

minimizer

monte_carlo

multifit_solver

multiset

n_tuples

ordinary_differential_equations
Numerical ODE solvers.

permutation

polynomial

qrng

ran_discrete
Given K discrete events with different probabilities P[k], produce a random value k consistent with its probability.

result

rng

series_acceleration

vector

vector_complex

wavelet_transforms