Struct rgsl::types::basis_spline::BSpLineDerivWorkspace

pub struct BSpLineDerivWorkspace {
    // some fields omitted
}

Methods

impl BSpLineDerivWorkspace

fn new(k: usize) -> Option<BSpLineDerivWorkspace>

This function allocates a workspace for computing the derivatives of a B-spline basis function of order k. The size of the workspace is O(2k2).

fn eval(&self, x: f64, nderiv: usize, dB: &MatrixF64, w: &BSpLineWorkspace) -> Value

This function evaluates all B-spline basis function derivatives of orders 0 through nderiv (inclusive) at the position x and stores them in the matrix dB. The (i,j)-th element of dB is djB_i(x)/dxj. The matrix dB must be of size n = nbreak + k - 2 by nderiv + 1. The value n may also be obtained by calling gsl_bspline_ncoeffs. Note that function evaluations are included as the zeroth order derivatives in dB. Computing all the basis function derivatives at once is more efficient than computing them individually, due to the nature of the defining recurrence relation.

fn eval_non_zero(&self, x: f64, nderiv: usize, dB: &MatrixF64, istart: &mut usize, iend: &mut usize, w: &BSpLineWorkspace) -> Value

This function evaluates all potentially nonzero B-spline basis function derivatives of orders 0 through nderiv (inclusive) at the position x and stores them in the matrix dB. The (i,j)-th element of dB is dj/dxj B_(istart+i)(x). The last row of dB contains dj/dxj B_(iend)(x). The matrix dB must be of size k by at least nderiv + 1. Note that function evaluations are included as the zeroth order derivatives in dB. By returning only the nonzero basis functions, this function allows quantities involving linear combinations of the B_i(x) and their derivatives to be computed without unnecessary terms.