Struct rgsl::types::eigen_symmetric_workspace::EigenGenHermWorkspace

pub struct EigenGenHermWorkspace {
    // some fields omitted
}

Methods

impl EigenGenHermWorkspace

fn new(n: usize) -> Option<EigenGenHermWorkspace>

This function allocates a workspace for computing eigenvalues of n-by-n complex generalized hermitian-definite eigensystems. The size of the workspace is O(3n).

fn genherm(&self, A: &MatrixComplexF64, B: &MatrixComplexF64, eval: &VectorF64) -> Value

This function computes the eigenvalues of the complex generalized hermitian-definite matrix pair (A, B), and stores them in eval, using the method outlined above. On output, B contains its Cholesky decomposition and A is destroyed.