Struct rgsl::types::fast_fourier_transforms::FftComplexWaveTable

pub struct FftComplexWaveTable {
    // some fields omitted
}

Methods

impl FftComplexWaveTable

fn new(n: usize) -> Option<FftComplexWaveTable>

This function prepares a trigonometric lookup table for a complex FFT of length n. The function returns a pointer to the newly allocated gsl_fft_complex_wavetable if no errors were detected, and a null pointer in the case of error. The length n is factorized into a product of subtransforms, and the factors and their trigonometric coefficients are stored in the wavetable. The trigonometric coefficients are computed using direct calls to sin and cos, for accuracy. Recursion relations could be used to compute the lookup table faster, but if an application performs many FFTs of the same length then this computation is a one-off overhead which does not affect the final throughput.

The wavetable structure can be used repeatedly for any transform of the same length. The table is not modified by calls to any of the other FFT functions. The same wavetable can be used for both forward and backward (or inverse) transforms of a given length.

fn factor<'r>(&'r mut self) -> &'r mut [usize]