Struct rgsl::types::integration::IntegrationQawoTable

pub struct IntegrationQawoTable {
    // some fields omitted
}

The QAWO algorithm is designed for integrands with an oscillatory factor, \sin(\omega x) or \cos(\omega x). In order to work efficiently the algorithm requires a table of Chebyshev moments which must be pre-computed with calls to the functions below.

Methods

impl IntegrationQawoTable

fn new(omega: f64, l: f64, sine: IntegrationQawo, n: usize) -> Option<IntegrationQawoTable>

This function allocates space for a gsl_integration_qawo_table struct and its associated workspace describing a sine or cosine weight function W(x) with the parameters (\omega, L),

W(x) = sin(omega x) W(x) = cos(omega x)

The parameter L must be the length of the interval over which the function will be integrated L = b - a. The choice of sine or cosine is made with the parameter sine which should be chosen from one of the two following symbolic values:

::Cosine ::IntegrationQawo::Sine

The gsl_integration_qawo_table is a table of the trigonometric coefficients required in the integration process. The parameter n determines the number of levels of coefficients that are computed. Each level corresponds to one bisection of the interval L, so that n levels are sufficient for subintervals down to the length L/2n. The integration routine gsl_integration_qawo returns the error ::Table if the number of levels is insufficient for the requested accuracy.

fn set(&self, omega: f64, l: f64, sine: IntegrationQawo) -> Value

This function changes the parameters omega, L and sine of the existing self workspace.

fn set_length(&self, l: f64) -> Value

This function allows the length parameter l of the self workspace to be changed.

fn qawo<T>(&self, f: function<T>, arg: &mut T, a: f64, epsabs: f64, epsrel: f64, limit: usize, workspace: &IntegrationWorkspace, result: &mut f64, abserr: &mut f64) -> Value

This function uses an adaptive algorithm to compute the integral of f over (a,b) with the weight function \sin(\omega x) or \cos(\omega x) defined by the table wf,

I = \int_ab dx f(x) sin(omega x) I = \int_ab dx f(x) cos(omega x)

The results are extrapolated using the epsilon-algorithm to accelerate the convergence of the integral. The function returns the final approximation from the extrapolation, result, and an estimate of the absolute error, abserr. The subintervals and their results are stored in the memory provided by workspace. The maximum number of subintervals is given by limit, which may not exceed the allocated size of the workspace.

Those subintervals with “large” widths d where d\omega > 4 are computed using a 25-point Clenshaw-Curtis integration rule, which handles the oscillatory behavior. Subintervals with a “small” widths where d\omega < 4 are computed using a 15-point Gauss-Kronrod integration.