Struct rgsl::types::n_tuples::NTuples

pub struct NTuples<T> {
    // some fields omitted
}

Methods

impl<T> NTuples<T>

fn create(filename: &str, data: &mut T) -> Option<NTuples<T>>

This function creates a new write-only ntuple file filename for ntuples of size size and returns a pointer to the newly created ntuple struct. Any existing file with the same name is truncated to zero length and overwritten. A pointer to memory for the current ntuple row ntuple_data must be supplied—this is used to copy ntuples in and out of the file.

fn open(filename: &str, data: &mut T) -> Option<NTuples<T>>

This function opens an existing ntuple file filename for reading and returns a pointer to a corresponding ntuple struct. The ntuples in the file must have size size. A pointer to memory for the current ntuple row ntuple_data must be supplied—this is used to copy ntuples in and out of the file.

fn write(&self) -> Value

This function writes the current ntuple ntuple->ntuple_data of size ntuple->size to the corresponding file.

fn bookdata(&self) -> Value

This function is a synonym for NTuples::write.

fn read(&self) -> Value

This function reads the current row of the ntuple file for ntuple and stores the values in ntuple->data.

fn project<U, V>(&self, h: &Histogram, value_func: value_function<T, U>, value_arg: &mut U, select_func: select_function<T, V>, select_arg: &mut V) -> Value