Struct rgsl::types::ordinary_differential_equations::ODEiv2StepType

pub struct ODEiv2StepType {
    // some fields omitted
}

Methods

impl ODEiv2StepType

fn rk2() -> ODEiv2StepType

Explicit embedded Runge-Kutta (2, 3) method.

fn rk4() -> ODEiv2StepType

Explicit 4th order (classical) Runge-Kutta. Error estimation is carried out by the step doubling method. For more efficient estimate of the error, use the embedded methods described below.

fn rk45() -> ODEiv2StepType

Explicit embedded Runge-Kutta-Fehlberg (4, 5) method. This method is a good general-purpose integrator.

fn rkck() -> ODEiv2StepType

Explicit embedded Runge-Kutta Cash-Karp (4, 5) method.

fn rk8pd() -> ODEiv2StepType

Explicit embedded Runge-Kutta Prince-Dormand (8, 9) method.

fn rk1imp() -> ODEiv2StepType

Implicit Gaussian first order Runge-Kutta. Also known as implicit Euler or backward Euler method. Error estimation is carried out by the step doubling method. This algorithm requires the Jacobian and access to the driver object via gsl_odeiv2_step_set_driver.

fn rk2imp() -> ODEiv2StepType

Implicit Gaussian second order Runge-Kutta. Also known as implicit mid-point rule. Error estimation is carried out by the step doubling method. This stepper requires the Jacobian and access to the driver object via gsl_odeiv2_step_set_driver.

fn rk4imp() -> ODEiv2StepType

Implicit Gaussian 4th order Runge-Kutta. Error estimation is carried out by the step doubling method. This algorithm requires the Jacobian and access to the driver object via gsl_odeiv2_step_set_driver.

fn bsimp() -> ODEiv2StepType

Implicit Bulirsch-Stoer method of Bader and Deuflhard. The method is generally suitable for stiff problems. This stepper requires the Jacobian.

fn msadams() -> ODEiv2StepType

A variable-coefficient linear multistep Adams method in Nordsieck form. This stepper uses explicit Adams-Bashforth (predictor) and implicit Adams-Moulton (corrector) methods in P(EC)m functional iteration mode. Method order varies dynamically between 1 and 12. This stepper requires the access to the driver object via gsl_odeiv2_step_set_driver.

fn msbdf() -> ODEiv2StepType

A variable-coefficient linear multistep backward differentiation formula (BDF) method in Nordsieck form. This stepper uses the explicit BDF formula as predictor and implicit BDF formula as corrector. A modified Newton iteration method is used to solve the system of non-linear equations. Method order varies dynamically between 1 and 5. The method is generally suitable for stiff problems. This stepper requires the Jacobian and the access to the driver object via gsl_odeiv2_step_set_driver.