Struct rgsl::types::ordinary_differential_equations::ODEiv2System

pub struct ODEiv2System<'a> {
    // some fields omitted
}

Description of a system of ODEs.

y' = f(t,y) = dydt(t, y)

The system is specified by giving the right-hand-side of the equation and possibly a jacobian function.

Some methods require the jacobian function, which calculates the matrix dfdy and the vector dfdt. The matrix dfdy conforms to the GSL standard, being a continuous range of floating point values, in row-order.

Methods

impl<'a> ODEiv2System<'a>

fn new(dimension: usize, function: &'a mut FnMut(f64, &[f64], &mut [f64]) -> GSLResult<()>) -> ODEiv2System<'a>

Returns a new ODEiv2System with a given dimension and right-hand side.

fn with_jacobian(dimension: usize, function: &'a mut FnMut(f64, &[f64], &mut [f64]) -> GSLResult<()>, jacobian: &'a mut FnMut(f64, &[f64], &mut [f64], &mut [f64]) -> GSLResult<()>) -> ODEiv2System<'a>

Returns a new ODEiv2System with a jacobian function provided.