Struct rgsl::types::result::ResultE10

pub struct ResultE10 {
    pub val: f64,
    pub err: f64,
    pub e10: i32,
}

In some cases, an overflow or underflow can be detected and handled by a function. In this case, it may be possible to return a scaling exponent as well as an error/value pair in order to save the result from exceeding the dynamic range of the built-in types.

Fields
val: f64

Contains the value.

err: f64

Contains an estimate of the absolute error in the value.

e10: i32

Exponent field such that the actual result is obtained as result * 10e10.

Methods

impl ResultE10

fn new() -> ResultE10