Function rgsl::types::rng::algorithms::mrg

pub fn mrg() -> RngType

This is a fifth-order multiple recursive generator by L’Ecuyer, Blouin and Coutre. Its sequence is,

x_n = (a_1 x_{n-1} + a_5 x_{n-5}) mod m

with a_1 = 107374182, a_2 = a_3 = a_4 = 0, a_5 = 104480 and m = 231 - 1.

The period of this generator is about 1046. It uses 5 words of state per generator. More information can be found in the following paper,

P. L’Ecuyer, F. Blouin, and R. Coutre, “A search for good multiple recursive random number generators”, ACM Transactions on Modeling and Computer Simulation 3, 87–98 (1993).