Struct rgsl::types::roots::RootFSolver

pub struct RootFSolver { /* fields omitted */ }

Methods

impl RootFSolver

pub fn new(t: &RootFSolverType) -> Option<RootFSolver>

This function returns a pointer to a newly allocated instance of a solver of type T.

If there is insufficient memory to create the solver then the function returns a null pointer and the error handler is invoked with an error code of Value::NoMemory.

pub fn set(&mut self, f: &mut RootFunction, x_lower: f64, x_upper: f64) -> Value

This function initializes, or reinitializes, an existing solver s to use the function f and the initial search interval [x lower, x upper].

pub fn iterate(&mut self) -> Value

The following function drives the iteration of each algorithm. Each function performs one iteration to update the state of any solver of the corresponding type. The same func- tion works for all solvers so that different methods can be substituted at runtime without modifications to the code.

This function performs a single iteration of the solver s. If the iteration encounters an unexpected problem then an error code will be returned.

The solver maintains a current best estimate of the root at all times. The bracketing solvers also keep track of the current best interval bounding the root.

pub fn name(&self) -> String

Returns the solver type name.

pub fn root(&self) -> f64

This function returns the current estimate of the root for the solver s.

pub fn x_lower(&self) -> f64

These functions return the current bracketing interval for the solver s.

pub fn x_upper(&self) -> f64

These functions return the current bracketing interval for the solver s.