Struct rgsl::types::siman::SimAnnealing

pub struct SimAnnealing<T: Clone> { /* fields omitted */ }

Methods

impl<T> SimAnnealing<T> where
T: Clone,

pub fn new(
x0_p: T,
ef: fn(_: &T) -> f64,
take_step: fn(_: &mut Rng, _: &mut T, _: f64),
distance: fn(_: &T, _: &T) -> f64,
print_pos: Option<fn(_: &T)>,
params: SimAnnealingParams
) -> SimAnnealing<T>

pub fn solve(&self, rng: &mut Rng) -> T

This function performs a simulated annealing search through a given space. The space is specified by providing the functions Ef and distance. The simulated annealing steps are generated using the random number generator rng and the function take_step.

The starting configuration of the system should be given by x0_p.

The params structure (described below) controls the run by providing the temperature schedule and other tunable parameters to the algorithm.

On exit the best result achieved during the search is returned. If the annealing process has been successful this should be a good approximation to the optimal point in the space.

If the argument print_pos is not None, a debugging log will be printed to stdout with the following columns: #-iter #-evals temperature position energy best_energy and the output of the function print position itself.

pub fn solve_many(&self, rng: &mut Rng) -> T

Like the function solve, but performs multiple runs and returns the best result.