Struct rgsl::types::vector_complex::VectorComplexF32

pub struct VectorComplexF32 { /* fields omitted */ }

Methods

impl VectorComplexF32

pub fn new(size: usize) -> Option<VectorComplexF32>

create a new VectorComplexF32 with all elements set to zero

pub fn from_slice(slice: &[ComplexF32]) -> Option<VectorComplexF32>

pub fn len(&self) -> usize

pub fn get(&self, i: usize) -> ComplexF32

This function returns the i-th element of a vector v. If i lies outside the allowed range of 0 to n-1 then the error handler is invoked and 0 is returned.

pub fn set(&mut self, i: usize, x: &ComplexF32) -> &VectorComplexF32

This function sets the value of the i-th element of a vector v to x. If i lies outside the allowed range of 0 to n-1 then the error handler is invoked.

pub fn set_all(&mut self, x: &ComplexF32) -> &VectorComplexF32

This function sets all the elements of the vector v to the value x.

pub fn set_zero(&mut self) -> &VectorComplexF32

This function sets all the elements of the vector v to zero.

pub fn set_basis(&mut self, i: usize) -> &VectorComplexF32

This function makes a basis vector by setting all the elements of the vector v to zero except for the i-th element which is set to one.

pub fn copy_from(&mut self, other: &VectorComplexF32) -> Value

This function copies the elements of the other vector into the self vector. The two vectors must have the same length.

pub fn copy_to(&self, other: &mut VectorComplexF32) -> Value

This function copies the elements of the self vector into the other vector. The two vectors must have the same length.

pub fn swap(&mut self, other: &mut VectorComplexF32) -> Value

This function exchanges the elements of the vectors by copying. The two vectors must have the same length.

pub fn swap_elements(&mut self, i: usize, j: usize) -> Value

This function exchanges the i-th and j-th elements of the vector v in-place.

pub fn reverse(&mut self) -> Value

This function reverses the order of the elements of the vector v.

pub fn add(&mut self, other: &VectorComplexF32) -> Value

This function adds the elements of the other vector to the elements of the self vector. The result a_i <- a_i + b_i is stored in self and other remains unchanged. The two vectors must have the same length.

pub fn sub(&mut self, other: &VectorComplexF32) -> Value

This function subtracts the elements of the self vector from the elements of the other vector. The result a_i <- a_i - b_i is stored in self and other remains unchanged. The two vectors must have the same length.

pub fn mul(&mut self, other: &VectorComplexF32) -> Value

This function multiplies the elements of the self vector a by the elements of the other vector. The result a_i <- a_i * b_i is stored in self and other remains unchanged. The two vectors must have the same length.

pub fn div(&mut self, other: &VectorComplexF32) -> Value

This function divides the elements of the self vector by the elements of the other vector. The result a_i <- a_i / b_i is stored in self and other remains unchanged. The two vectors must have the same length.

pub fn scale(&mut self, x: &ComplexF32) -> Value

This function multiplies the elements of the self vector by the constant factor x. The result a_i <- a_i is stored in self.

pub fn add_constant(&mut self, x: &ComplexF32) -> Value

This function adds the constant value x to the elements of the self vector. The result a_i <- a_i + x is stored in self.

pub fn is_null(&self) -> bool

This function returns true if all the elements of the self vector are equal to 0.

pub fn is_pos(&self) -> bool

This function returns true if all the elements of the self vector are stricly positive.

pub fn is_neg(&self) -> bool

This function returns true if all the elements of the self vector are stricly negative.

pub fn is_non_neg(&self) -> bool

This function returns true if all the elements of the self vector are stricly non-negative.

pub fn equal(&self, other: &VectorComplexF32) -> bool

pub fn clone(&self) -> Option<VectorComplexF32>