Struct rgsl::types::wavelet_transforms::WaveletWorkspace

pub struct WaveletWorkspace {
    // some fields omitted
}

The WaveletWorkspace structure contains scratch space of the same size as the input data and is used to hold intermediate results during the transform.

Methods

impl WaveletWorkspace

fn new(n: usize) -> Option<WaveletWorkspace>

This function allocates a workspace for the discrete wavelet transform. To perform a one-dimensional transform on n elements, a workspace of size n must be provided. For two-dimensional transforms of n-by-n matrices it is sufficient to allocate a workspace of size n, since the transform operates on individual rows and columns. A null pointer is returned if insufficient memory is available.