Module rgsl::zeta::riemann_mins_one

For large positive argument, the Riemann zeta function approaches one. In this region the fractional part is interesting, and therefore we need a function to evaluate it explicitly.

Functions

zetam1
This routine computes \zeta(s) - 1 for arbitrary s, s \ne 1.

zetam1_e
This routine computes \zeta(s) - 1 for arbitrary s, s \ne 1.

zetam1_int
This routine computes the Riemann zeta function \zeta(n) for integer n, n \ne 1.

zetam1_int_e
This routine computes the Riemann zeta function \zeta(n) for integer n, n \ne 1.