Module rgsl::bessel

Functions

I0
This routine computes the regular modified cylindrical Bessel function of zeroth order, I_0(x)

I0_e
This routine computes the regular modified cylindrical Bessel function of zeroth order, I_0(x)

I0_scaled
This routine computes the scaled regular modified cylindrical Bessel function of zeroth order \exp(-|x|) I_0(x).

I0_scaled_e
This routine computes the scaled regular modified cylindrical Bessel function of zeroth order \exp(-|x|) I_0(x).

I1
This routine computes the regular modified cylindrical Bessel function of first order, I_1(x).

I1_e
This routine computes the regular modified cylindrical Bessel function of first order, I_1(x).

I1_scaled
This routine computes the scaled regular modified cylindrical Bessel function of first order \exp(-|x|) I_1(x).

I1_scaled_e
This routine computes the scaled regular modified cylindrical Bessel function of first order \exp(-|x|) I_1(x).

In
This routine computes the regular modified cylindrical Bessel function of order n, I_n(x).

In_array
This routine computes the values of the regular modified cylindrical Bessel functions I_n(x) for n from nmin to nmax inclusive, storing the results in the array result_array. The start of the range nmin must be positive or zero. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

In_e
This routine computes the regular modified cylindrical Bessel function of order n, I_n(x).

In_scaled
This routine computes the scaled regular modified cylindrical Bessel function of order n, \exp(-|x|) I_n(x)

In_scaled_array
This routine computes the values of the scaled regular cylindrical Bessel functions \exp(-|x|) I_n(x) for n from nmin to nmax inclusive, storing the results in the array result_array. The start of the range nmin must be positive or zero. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

In_scaled_e
This routine computes the scaled regular modified cylindrical Bessel function of order n, \exp(-|x|) I_n(x)

Inu
This routine computes the regular modified Bessel function of fractional order \nu, I_\nu(x) for x>0, \nu>0.

Inu_e
This routine computes the regular modified Bessel function of fractional order \nu, I_\nu(x) for x>0, \nu>0.

Inu_scaled
This routine computes the scaled regular modified Bessel function of fractional order \nu, \exp(-|x|)I_\nu(x) for x>0, \nu>0.

Inu_scaled_e
This routine computes the scaled regular modified Bessel function of fractional order \nu, \exp(-|x|)I_\nu(x) for x>0, \nu>0.

J0
This routine computes the regular cylindrical Bessel function of zeroth order, J_0(x).

J0_e
This routine computes the regular cylindrical Bessel function of zeroth order, J_0(x).

J1
This routine computes the regular cylindrical Bessel function of first order, J_1(x).

J1_e
This routine computes the regular cylindrical Bessel function of first order, J_1(x).

Jn
This routine computes the regular cylindrical Bessel function of order n, J_n(x).

Jn_array
This routine computes the values of the regular cylindrical Bessel functions J_n(x) for n from nmin to nmax inclusive, storing the results in the array result_array. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

Jn_e
This routine computes the regular cylindrical Bessel function of order n, J_n(x).

Jnu
This routine computes the regular cylindrical Bessel function of fractional order \nu, J_\nu(x).

Jnu_e
This routine computes the regular cylindrical Bessel function of fractional order \nu, J_\nu(x).

K0
This routine computes the irregular modified cylindrical Bessel function of zeroth order, K_0(x), for x > 0.

K0_e
This routine computes the irregular modified cylindrical Bessel function of zeroth order, K_0(x), for x > 0.

K0_scaled
This routine computes the scaled irregular modified cylindrical Bessel function of zeroth order \exp(x) K_0(x) for x>0.

K0_scaled_e
This routine computes the scaled irregular modified cylindrical Bessel function of zeroth order \exp(x) K_0(x) for x>0.

K1
This routine computes the irregular modified cylindrical Bessel function of first order, K_1(x), for x > 0.

K1_e
This routine computes the irregular modified cylindrical Bessel function of first order, K_1(x), for x > 0.

K1_scaled
This routine computes the scaled irregular modified cylindrical Bessel function of first order \exp(x) K_1(x) for x>0.

K1_scaled_e
This routine computes the scaled irregular modified cylindrical Bessel function of first order \exp(x) K_1(x) for x>0.

Kn
This routine computes the irregular modified cylindrical Bessel function of order n, K_n(x), for x > 0.

Kn_array
This routine computes the values of the irregular modified cylindrical Bessel functions K_n(x) for n from nmin to nmax inclusive, storing the results in the array result_array. The start of the range nmin must be positive or zero. The domain of the function is x>0. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

Kn_e
This routine computes the irregular modified cylindrical Bessel function of order n, K_n(x), for x > 0.

Kn_scaled
This routine computes the scaled irregular modified cylindrical Bessel function of order n, \exp(x) K_n(x), for x>0.

Kn_scaled_array
This routine computes the values of the scaled irregular cylindrical Bessel functions \exp(x) K_n(x) for n from nmin to nmax inclusive, storing the results in the array result_array. The start of the range nmin must be positive or zero. The domain of the function is x>0. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

Kn_scaled_e
This routine computes the scaled irregular modified cylindrical Bessel function of order n, \exp(x) K_n(x), for x>0.

Knu
This routine computes the irregular modified Bessel function of fractional order \nu, K_\nu(x) for x>0, \nu>0.

Knu_e
This routine computes the irregular modified Bessel function of fractional order \nu, K_\nu(x) for x>0, \nu>0.

Knu_scaled
This routine computes the scaled irregular modified Bessel function of fractional order \nu, \exp(+|x|) K_\nu(x) for x>0, \nu>0.

Knu_scaled_e
This routine computes the scaled irregular modified Bessel function of fractional order \nu, \exp(+|x|) K_\nu(x) for x>0, \nu>0.

Y0
This routine computes the irregular cylindrical Bessel function of zeroth order, Y_0(x), for x>0.

Y0_e
This routine computes the irregular cylindrical Bessel function of zeroth order, Y_0(x), for x>0.

Y1
This routine computes the irregular cylindrical Bessel function of first order, Y_1(x), for x>0.

Y1_e
This routine computes the irregular cylindrical Bessel function of first order, Y_1(x), for x>0.

Yn
This routine computes the irregular cylindrical Bessel function of order n, Y_n(x), for x>0.

Yn_array
This routine computes the values of the irregular cylindrical Bessel functions Y_n(x) for n from nmin to nmax inclusive, storing the results in the array result_array. The domain of the function is x>0. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

Yn_e
This routine computes the irregular cylindrical Bessel function of order n, Y_n(x), for x>0.

Ynu
This routine computes the irregular cylindrical Bessel function of fractional order \nu, Y_\nu(x).

Ynu_e
This routine computes the irregular cylindrical Bessel function of fractional order \nu, Y_\nu(x).

i0_scaled
This routine computes the scaled regular modified spherical Bessel function of zeroth order, \exp(-|x|) i_0(x).

i0_scaled_e
This routine computes the scaled regular modified spherical Bessel function of zeroth order, \exp(-|x|) i_0(x).

i1_scaled
This routine computes the scaled regular modified spherical Bessel function of first order, \exp(-|x|) i_1(x).

i1_scaled_e
This routine computes the scaled regular modified spherical Bessel function of first order, \exp(-|x|) i_1(x).

i2_scaled
This routine computes the scaled regular modified spherical Bessel function of second order, \exp(-|x|) i_2(x)

i2_scaled_e
This routine computes the scaled regular modified spherical Bessel function of second order, \exp(-|x|) i_2(x)

il_scaled
This routine computes the scaled regular modified spherical Bessel function of order l, \exp(-|x|) i_l(x)

il_scaled_array
This routine computes the values of the scaled regular modified cylindrical Bessel functions \exp(-|x|) i_l(x) for l from 0 to lmax inclusive for lmax >= 0, storing the results in the array result_array. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

il_scaled_e
This routine computes the scaled regular modified spherical Bessel function of order l, \exp(-|x|) i_l(x)

j0
This routine computes the regular spherical Bessel function of zeroth order, j_0(x) = \sin(x)/x.

j0_e
This routine computes the regular spherical Bessel function of zeroth order, j_0(x) = \sin(x)/x.

j1
This routine computes the regular spherical Bessel function of first order, j_1(x) = (\sin(x)/x - \cos(x))/x.

j1_e
This routine computes the regular spherical Bessel function of first order, j_1(x) = (\sin(x)/x - \cos(x))/x.

j2
This routine computes the regular spherical Bessel function of second order, j_2(x) = ((3/x^2 - 1)\sin(x) - 3\cos(x)/x)/x.

j2_e
This routine computes the regular spherical Bessel function of second order, j_2(x) = ((3/x^2 - 1)\sin(x) - 3\cos(x)/x)/x.

jl
This routine computes the regular spherical Bessel function of order l, j_l(x), for l >= 0 and x >= 0.

jl_array
This routine computes the values of the regular spherical Bessel functions j_l(x) for l from 0 to lmax inclusive for lmax >= 0 and x >= 0, storing the results in the array result_array. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

jl_e
This routine computes the regular spherical Bessel function of order l, j_l(x), for l >= 0 and x >= 0.

jl_steed_array
This routine uses Steed’s method to compute the values of the regular spherical Bessel functions j_l(x) for l from 0 to lmax inclusive for lmax >= 0 and x >= 0, storing the results in the array result_array. The Steed/Barnett algorithm is described in Comp. Phys. Comm. 21, 297 (1981). Steed’s method is more stable than the recurrence used in the other functions but is also slower.

k0_scaled
The irregular modified spherical Bessel functions k_l(x) are related to the irregular modified Bessel functions of fractional order, k_l(x) = \sqrt{\pi/(2x)} K_{l+1/2}(x). This routine computes the scaled irregular modified spherical Bessel function of zeroth order, \exp(x) k_0(x), for x>0.

k0_scaled_e
The irregular modified spherical Bessel functions k_l(x) are related to the irregular modified Bessel functions of fractional order, k_l(x) = \sqrt{\pi/(2x)} K_{l+1/2}(x). This routine computes the scaled irregular modified spherical Bessel function of zeroth order, \exp(x) k_0(x), for x>0.

k1_scaled
This routine computes the scaled irregular modified spherical Bessel function of first order, \exp(x) k_1(x), for x>0.

k1_scaled_e
This routine computes the scaled irregular modified spherical Bessel function of first order, \exp(x) k_1(x), for x>0.

k2_scaled
This routine computes the scaled irregular modified spherical Bessel function of second order, \exp(x) k_2(x), for x>0.

k2_scaled_e
This routine computes the scaled irregular modified spherical Bessel function of second order, \exp(x) k_2(x), for x>0.

kl_scaled
This routine computes the scaled irregular modified spherical Bessel function of order l, \exp(x) k_l(x), for x>0.

kl_scaled_array
This routine computes the values of the scaled irregular modified spherical Bessel functions \exp(x) k_l(x) for l from 0 to lmax inclusive for lmax >= 0 and x>0, storing the results in the array result_array. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

kl_scaled_e
This routine computes the scaled irregular modified spherical Bessel function of order l, \exp(x) k_l(x), for x>0.

lnKnu
This routine computes the logarithm of the irregular modified Bessel function of fractional order \nu, \ln(K_\nu(x)) for x>0, \nu>0.

lnKnu_e
This routine computes the logarithm of the irregular modified Bessel function of fractional order \nu, \ln(K_\nu(x)) for x>0, \nu>0.

sequence_Jnu
This function computes the regular cylindrical Bessel function of fractional order \nu, J_\nu(x), evaluated at a series of x values. The array v of length size contains the x values. They are assumed to be strictly ordered and positive. The array is over-written with the values of J_\nu(x_i).

y0
This routine computes the irregular spherical Bessel function of zeroth order, y_0(x) = -\cos(x)/x.

y0_e
This routine computes the irregular spherical Bessel function of zeroth order, y_0(x) = -\cos(x)/x.

y1
This routine computes the irregular spherical Bessel function of first order, y_1(x) = -(\cos(x)/x + \sin(x))/x.

y1_e
This routine computes the irregular spherical Bessel function of first order, y_1(x) = -(\cos(x)/x + \sin(x))/x.

y2
This routine computes the irregular spherical Bessel function of second order, y_2(x) = (-3/x^3 + 1/x)\cos(x) - (3/x^2)\sin(x).

y2_e
This routine computes the irregular spherical Bessel function of second order, y_2(x) = (-3/x^3 + 1/x)\cos(x) - (3/x^2)\sin(x).

yl
This routine computes the irregular spherical Bessel function of order l, y_l(x), for l >= 0.

yl_array
This routine computes the values of the irregular spherical Bessel functions y_l(x) for l from 0 to lmax inclusive for lmax >= 0, storing the results in the array result_array. The values are computed using recurrence relations for efficiency, and therefore may differ slightly from the exact values.

yl_e
This routine computes the irregular spherical Bessel function of order l, y_l(x), for l >= 0.

zero_J0
This routine computes the location of the s-th positive zero of the Bessel function J_0(x).

zero_J0_e
This routine computes the location of the s-th positive zero of the Bessel function J_0(x).

zero_J1
This routine computes the location of the s-th positive zero of the Bessel function J_1(x).

zero_J1_e
This routine computes the location of the s-th positive zero of the Bessel function J_1(x).

zero_Jnu
This routine computes the location of the s-th positive zero of the Bessel function J_\nu(x). The current implementation does not support negative values of nu.

zero_Jnu_e
This routine computes the location of the s-th positive zero of the Bessel function J_\nu(x). The current implementation does not support negative values of nu.