Module rgsl::blas::level1

Functions

caxpy

This function computes the sum y = \alpha x + y for the vectors x and y.

ccopy

This function copy the elements of the vector x into the vector y.

cdotc

This function computes the complex conjugate scalar product x^H y for the vectors x and y, returning the result in dotc.

cdotu

This function computes the complex scalar product x^T y for the vectors x and y, returning the result in dotu.

cscal

This function rescales the vector x by the multiplicative factor alpha.

csscal

This function rescales the vector x by the multiplicative factor alpha.

cswap

This function exchanges the elements of the vectors x and y.

dasum

This function computes the absolute sum \sum |x_i| of the elements of the vector x.

daxpy

This function computes the sum y = \alpha x + y for the vectors x and y.

dcopy

This function copy the elements of the vector x into the vector y.

ddot

This function computes the scalar product x^T y for the vectors x and y, returning the result in result.

dnrm2

This function computes the Euclidean norm ||x||_2 = \sqrt {\sum x_i^2} of the vector x.

drot

This function applies a Givens rotation (x’, y’) = (c x + s y, -s x + c y) to the vectors x, y.

drotg

This function computes a Givens rotation (c,s) which zeroes the vector (a,b),

drotm

This function applies a modified Givens transformation.

drotmg

This function computes a modified Givens transformation. The modified Givens transformation is defined in the original Level-1 BLAS specification, given in the references.

dscal

This function rescales the vector x by the multiplicative factor alpha.

dsdot

This function computes the scalar product x^T y for the vectors x and y, returning the result in result.

dswap

This function exchanges the elements of the vectors x and y.

dzasum

This function computes the sum of the magnitudes of the real and imaginary parts of the complex vector x, \sum |\Re(x_i)| + |\Im(x_i)|.

dznrm2

This function computes the Euclidean norm of the complex vector x,

icamax

This function returns the index of the largest element of the vector x. The largest element is determined by its absolute magnitude for real vectors and by the sum of the magnitudes of the real and imaginary parts |\Re(x_i)| + |\Im(x_i)| for complex vectors. If the largest value occurs several times then the index of the first occurrence is returned.

idamax

This function returns the index of the largest element of the vector x. The largest element is determined by its absolute magnitude for real vectors and by the sum of the magnitudes of the real and imaginary parts |\Re(x_i)| + |\Im(x_i)| for complex vectors. If the largest value occurs several times then the index of the first occurrence is returned.

isamax

This function returns the index of the largest element of the vector x. The largest element is determined by its absolute magnitude for real vectors and by the sum of the magnitudes of the real and imaginary parts |\Re(x_i)| + |\Im(x_i)| for complex vectors. If the largest value occurs several times then the index of the first occurrence is returned.

izamax

This function returns the index of the largest element of the vector x. The largest element is determined by its absolute magnitude for real vectors and by the sum of the magnitudes of the real and imaginary parts |\Re(x_i)| + |\Im(x_i)| for complex vectors. If the largest value occurs several times then the index of the first occurrence is returned.

sasum

This function computes the absolute sum \sum |x_i| of the elements of the vector x.

saxpy

This function computes the sum y = \alpha x + y for the vectors x and y.

scasum

This function computes the sum of the magnitudes of the real and imaginary parts of the complex vector x, \sum |\Re(x_i)| + |\Im(x_i)|.

scnrm2

This function computes the Euclidean norm of the complex vector x,

scopy

This function copy the elements of the vector x into the vector y.

sdot

This function computes the scalar product x^T y for the vectors x and y, returning the result in result.

sdsdot

This function computes the sum \alpha + x^T y for the vectors x and y, returning the result in result.

snrm2

This function computes the Euclidean norm ||x||_2 = \sqrt {\sum x_i^2} of the vector x.

srot

This function applies a Givens rotation (x’, y’) = (c x + s y, -s x + c y) to the vectors x, y.

srotg

This function computes a Givens rotation (c,s) which zeroes the vector (a,b),

srotm

This function applies a modified Givens transformation.

srotmg

This function computes a modified Givens transformation. The modified Givens transformation is defined in the original Level-1 BLAS specification, given in the references.

sscal

This function rescales the vector x by the multiplicative factor alpha.

sswap

This function exchanges the elements of the vectors x and y.

zaxpy

This function computes the sum y = \alpha x + y for the vectors x and y.

zcopy

This function copy the elements of the vector x into the vector y.

zdotc

This function computes the complex conjugate scalar product x^H y for the vectors x and y, returning the result in dotc.

zdotu

This function computes the complex scalar product x^T y for the vectors x and y, returning the result in dotu.

zdscal

This function rescales the vector x by the multiplicative factor alpha.

zscal

This function rescales the vector x by the multiplicative factor alpha.

zswap

This function exchanges the elements of the vectors x and y.