Module rgsl::cblas

Modules

level1

level2

level3

Structs

Index

Enums

Diag

Order
Indicates whether a matrix is in Row Major or Column Major order. Row major order is the native order for C programs, while Column major order is native for Fortran.

Side
Used to indicate the order of a matrix-matrix multiply.

Transpose
Used to represent transpose operations on a matrix.

Uplo
Used to indicate which part of a symmetric matrix to use.