Function rgsl::coulomb::wave_FG_e

pub fn wave_FG_e(
    eta: f64, 
    x: f64, 
    L_F: f64, 
    k: i32, 
    exp_F: &mut f64, 
    exp_G: &mut f64
) -> (Value, Result, Result, Result, Result)

This function computes the Coulomb wave functions F_L(\eta,x), G_{L-k}(\eta,x) and their derivatives F'L(\eta,x), G'{L-k}(\eta,x) with respect to x. The parameters are restricted to L, L-k > -1/2, x > 0 and integer k. Note that L itself is not restricted to being an integer. The results are stored in the parameters F, G for the function values and Fp, Gp for the derivative values. If an overflow occurs, GSL_EOVRFLW is returned and scaling exponents are stored in the modifiable parameters exp_F, exp_G.