Trait rgsl::elementary::Elementary

pub trait Elementary {
    fn log1p(&self) -> Self;

    fn expm1(&self) -> Self;

    fn hypot(&self, y: f64) -> Self;

    fn hypot3(&self, y: f64, z: f64) -> Self;

    fn acosh(&self) -> Self;

    fn asinh(&self) -> Self;

    fn atanh(&self) -> Self;

    fn ldexp(&self, e: i32) -> Self;

    fn frexp(&self, e: &mut i32) -> Self;
}

Required methods

fn log1p(&self) -> Self

This function computes the value of log(1+x) in a way that is accurate for small x. It provides an alternative to the BSD math function log1p(x).

fn expm1(&self) -> Self

This function computes the value of exp(x)-1 in a way that is accurate for small x. It provides an alternative to the BSD math function expm1(x).

fn hypot(&self, y: f64) -> Self

This function computes the value of sqrt{x^2 + y^2} in a way that avoids overflow. It provides an alternative to the BSD math function hypot(x,y).

fn hypot3(&self, y: f64, z: f64) -> Self

This function computes the value of sqrt{x^2 + y^2 + z^2} in a way that avoids overflow.

fn acosh(&self) -> Self

This function computes the value of arccosh(x). It provides an alternative to the standard math function acosh(x).

fn asinh(&self) -> Self

This function computes the value of arcsinh(x). It provides an alternative to the standard math function asinh(x).

fn atanh(&self) -> Self

This function computes the value of arctanh(x). It provides an alternative to the standard math function atanh(x).

fn ldexp(&self, e: i32) -> Self

This function computes the value of x * 2^e. It provides an alternative to the standard math function ldexp(x,e).

fn frexp(&self, e: &mut i32) -> Self

This function splits the number x into its normalized fraction f and exponent e, such that x = f * 2^e and 0.5 <= f < 1. The function returns f and stores the exponent in e. If x is zero, both f and e are set to zero. This function provides an alternative to the standard math function frexp(x, e).

Loading content...