Module rgsl::elliptic::carlson

The Carlson symmetric forms of elliptical integrals RC(x,y), RD(x,y,z), RF(x,y,z) and RJ(x,y,z,p) are defined by,

RC(x,y) = 1/2 \int_0^\infty dt (t+x)^(-1/2) (t+y)^(-1)

RD(x,y,z) = 3/2 \int_0^\infty dt (t+x)^(-1/2) (t+y)^(-1/2) (t+z)^(-3/2)

RF(x,y,z) = 1/2 \int_0^\infty dt (t+x)^(-1/2) (t+y)^(-1/2) (t+z)^(-1/2)

RJ(x,y,z,p) = 3/2 \int_0^\infty dt (t+x)^(-1/2) (t+y)^(-1/2) (t+z)^(-1/2) (t+p)^(-1)

Functions

ellint_RC

This routine computes the incomplete elliptic integral RC(x,y) to the accuracy specified by the mode variable mode.

ellint_RC_e

This routine computes the incomplete elliptic integral RC(x,y) to the accuracy specified by the mode variable mode.

ellint_RD

This routine computes the incomplete elliptic integral RD(x,y,z) to the accuracy specified by the mode variable mode.

ellint_RD_e

This routine computes the incomplete elliptic integral RD(x,y,z) to the accuracy specified by the mode variable mode.

ellint_RF

This routine computes the incomplete elliptic integral RF(x,y,z) to the accuracy specified by the mode variable mode.

ellint_RF_e

This routine computes the incomplete elliptic integral RF(x,y,z) to the accuracy specified by the mode variable mode.

ellint_RJ

This routine computes the incomplete elliptic integral RJ(x,y,z,p) to the accuracy specified by the mode variable mode.

ellint_RJ_e

This routine computes the incomplete elliptic integral RJ(x,y,z,p) to the accuracy specified by the mode variable mode.