Module rgsl::fermi_dirac::complete_integrals

The complete Fermi-Dirac integral F_j(x) is given by,

F_j(x) := (1/\Gamma(j+1)) \int_0^\infty dt (t^j / (\exp(t-x) + 1))

Note that the Fermi-Dirac integral is sometimes defined without the normalisation factor in other texts.

Functions

fermi_dirac_0
This routine computes the complete Fermi-Dirac integral with an index of 0. This integral is given by F_0(x) = \ln(1 + e^x).

fermi_dirac_0_e
This routine computes the complete Fermi-Dirac integral with an index of 0. This integral is given by F_0(x) = \ln(1 + e^x).

fermi_dirac_1
This routine computes the complete Fermi-Dirac integral with an index of 1, F_1(x) = \int_0^\infty dt (t /(\exp(t-x)+1)).

fermi_dirac_1_e
This routine computes the complete Fermi-Dirac integral with an index of 1, F_1(x) = \int_0^\infty dt (t /(\exp(t-x)+1)).

fermi_dirac_2
This routine computes the complete Fermi-Dirac integral with an index of 2, F_2(x) = (1/2) \int_0^\infty dt (t^2 /(\exp(t-x)+1)).

fermi_dirac_2_e
This routine computes the complete Fermi-Dirac integral with an index of 2, F_2(x) = (1/2) \int_0^\infty dt (t^2 /(\exp(t-x)+1)).

fermi_dirac_3half
This routine computes the complete Fermi-Dirac integral F_{3/2}(x).

fermi_dirac_3half_e
This routine computes the complete Fermi-Dirac integral F_{3/2}(x).

fermi_dirac_half
This routine computes the complete Fermi-Dirac integral F_{1/2}(x).

fermi_dirac_half_e
This routine computes the complete Fermi-Dirac integral F_{1/2}(x).

fermi_dirac_int
This routine computes the complete Fermi-Dirac integral with an integer index of j, F_j(x) = (1/\Gamma(j+1)) \int_0^\infty dt (t^j /(\exp(t-x)+1)).

fermi_dirac_int_e
This routine computes the complete Fermi-Dirac integral with an integer index of j, F_j(x) = (1/\Gamma(j+1)) \int_0^\infty dt (t^j /(\exp(t-x)+1)).

fermi_dirac_m1
This routine computes the complete Fermi-Dirac integral with an index of -1. This integral is given by F_{-1}(x) = e^x / (1 + e^x).

fermi_dirac_m1_e
This routine computes the complete Fermi-Dirac integral with an index of -1. This integral is given by F_{-1}(x) = e^x / (1 + e^x).

fermi_dirac_mhalf
This routine computes the complete Fermi-Dirac integral F_{-1/2}(x).

fermi_dirac_mhalf_e
This routine computes the complete Fermi-Dirac integral F_{-1/2}(x).