Module rgsl::hypergeometric

Hypergeometric functions are described in Abramowitz & Stegun, Chapters 13 and 15.

Functions

hyperg_0F1
This routine computes the hypergeometric function 0F1(c,x).

hyperg_0F1_e
This routine computes the hypergeometric function 0F1(c,x).

hyperg_1F1_int
This routine computes the confluent hypergeometric function 1F1(m,n,x) = M(m,n,x) for integer parameters m, n.

hyperg_1F1_int_e
This routine computes the confluent hypergeometric function 1F1(m,n,x) = M(m,n,x) for integer parameters m, n.

hyperg_1F1
This routine computes the confluent hypergeometric function 1F1(a,b,x) = M(a,b,x) for general parameters a, b.

hyperg_1F1_e
This routine computes the confluent hypergeometric function 1F1(a,b,x) = M(a,b,x) for general parameters a, b.

hyperg_1F1_U_int
This routine computes the confluent hypergeometric function U(m,n,x) for integer parameters m, n.

hyperg_1F1_U_int_e
This routine computes the confluent hypergeometric function U(m,n,x) for integer parameters m, n.

hyperg_1F1_U_int_e10_e
This routine computes the confluent hypergeometric function U(m,n,x) for integer parameters m, n using the [ResultE10](types/result/struct.ResultE10.html) type to return a result with extended range.

hyperg_2F0
This routine computes the hypergeometric function 2F0(a,b,x). The series representation is a divergent hypergeometric series. However, for x < 0 we have 2F0(a,b,x) = (-1/x)^a U(a,1+a-b,-1/x)

hyperg_2F0_e
This routine computes the hypergeometric function 2F0(a,b,x). The series representation is a divergent hypergeometric series. However, for x < 0 we have 2F0(a,b,x) = (-1/x)^a U(a,1+a-b,-1/x)

hyperg_2F1
This routine computes the Gauss hypergeometric function 2F1(a,b,c,x) = F(a,b,c,x) for |x| < 1.

hyperg_2F1_e
This routine computes the Gauss hypergeometric function 2F1(a,b,c,x) = F(a,b,c,x) for |x| < 1.

hyperg_2F1_conj
This routine computes the Gauss hypergeometric function 2F1(a_R + i a_I, a_R - i a_I, c, x) with complex parameters for |x| < 1.

hyperg_2F1_conj_e
This routine computes the Gauss hypergeometric function 2F1(a_R + i a_I, a_R - i a_I, c, x) with complex parameters for |x| < 1.

hyperg_2F1_renorm
This routine computes the renormalized Gauss hypergeometric function 2F1(a,b,c,x) / \Gamma(c) for |x| < 1.

hyperg_2F1_renorm_e
This routine computes the renormalized Gauss hypergeometric function 2F1(a,b,c,x) / \Gamma(c) for |x| < 1.

hyperg_2F1_conj_renorm
This routine computes the renormalized Gauss hypergeometric function 2F1(a_R + i a_I, a_R - i a_I, c, x) / \Gamma(c) for |x| < 1.

hyperg_2F1_conj_renorm_e
This routine computes the renormalized Gauss hypergeometric function 2F1(a_R + i a_I, a_R - i a_I, c, x) / \Gamma(c) for |x| < 1.

hyperg_U
This routine computes the confluent hypergeometric function U(a,b,x).

hyperg_U_e
This routine computes the confluent hypergeometric function U(a,b,x).

hyperg_U_e10_e
This routine computes the confluent hypergeometric function U(a,b,x) using the [ResultE10](types/result/struct.ResultE10.html) type to return a result with extended range.