Module rgsl::lambert_w

Lambert’s W functions, W(x), are defined to be solutions of the equation W(x) \exp(W(x)) = x. This function has multiple branches for x < 0; however, it has only two real-valued branches. We define W_0(x) to be the principal branch, where W > -1 for x < 0, and W_{-1}(x) to be the other real branch, where W < -1 for x < 0.

Functions

lambert_W0
This computes the principal branch of the Lambert W function, W_0(x).

lambert_W0_e
This computes the principal branch of the Lambert W function, W_0(x).

lambert_Wm1
This computes the secondary real-valued branch of the Lambert W function, W_{-1}(x).

lambert_Wm1_e
This computes the secondary real-valued branch of the Lambert W function, W_{-1}(x).