Module rgsl::legendre::conical

The Conical Functions P^\mu_{-(1/2)+i\lambda}(x) and Q^\mu_{-(1/2)+i\lambda} are described in Abramowitz & Stegun, Section 8.12.

Functions

_0
This routine computes the conical function P^0_{-1/2 + i \lambda}(x) for x > -1.

_0_e
This routine computes the conical function P^0_{-1/2 + i \lambda}(x) for x > -1.

_1
This routine computes the conical function P^1_{-1/2 + i \lambda}(x) for x > -1.

_1_e
This routine computes the conical function P^1_{-1/2 + i \lambda}(x) for x > -1.

cyl_reg
This routine computes the Regular Cylindrical Conical Function P^{-m}_{-1/2 + i \lambda}(x) for x > -1, m >= -1.

cyl_reg_e
This routine computes the Regular Cylindrical Conical Function P^{-m}_{-1/2 + i \lambda}(x) for x > -1, m >= -1.

half
This routine computes the irregular Spherical Conical Function P^{1/2}_{-1/2 + i \lambda}(x) for x > -1.

half_e
This routine computes the irregular Spherical Conical Function P^{1/2}_{-1/2 + i \lambda}(x) for x > -1.

mhalf
This routine computes the regular Spherical Conical Function P^{-1/2}_{-1/2 + i \lambda}(x) for x > -1.

mhalf_e
This routine computes the regular Spherical Conical Function P^{-1/2}_{-1/2 + i \lambda}(x) for x > -1.

sph_reg
This routine computes the Regular Spherical Conical Function P^{-1/2-l}_{-1/2 + i \lambda}(x) for x > -1, l >= -1.

sph_reg_e
This routine computes the Regular Spherical Conical Function P^{-1/2-l}_{-1/2 + i \lambda}(x) for x > -1, l >= -1.