Module rgsl::legendre::polynomials

Functions

legendre_P1

This function evaluates the Legendre polynomials P_l(x) using explicit representations for l=1, 2, 3.

legendre_P1_e

This function evaluates the Legendre polynomials P_l(x) using explicit representations for l=1, 2, 3.

legendre_P2

This function evaluates the Legendre polynomials P_l(x) using explicit representations for l=1, 2, 3.

legendre_P2_e

This function evaluates the Legendre polynomials P_l(x) using explicit representations for l=1, 2, 3.

legendre_P3

This function evaluates the Legendre polynomials P_l(x) using explicit representations for l=1, 2, 3.

legendre_P3_e

This function evaluates the Legendre polynomials P_l(x) using explicit representations for l=1, 2, 3.

legendre_Pl

This function evaluates the Legendre polynomial P_l(x) for a specific value of l, x subject to l >= 0, |x| <= 1

legendre_Pl_array

This function computes arrays of Legendre polynomials P_l(x) and derivatives dP_l(x)/dx, for l = 0, \dots, lmax, |x| <= 1

legendre_Pl_deriv_array

This function computes arrays of Legendre polynomials P_l(x) and derivatives dP_l(x)/dx, for l = 0, \dots, lmax, |x| <= 1

legendre_Pl_e

This function evaluates the Legendre polynomial P_l(x) for a specific value of l, x subject to l >= 0, |x| <= 1

legendre_Q0

This function computes the Legendre function Q_0(x) for x > -1, x != 1

legendre_Q0_e

This function computes the Legendre function Q_0(x) for x > -1, x != 1

legendre_Q1

This function computes the Legendre function Q_0(x) for x > -1, x != 1.

legendre_Q1_e

This function computes the Legendre function Q_0(x) for x > -1, x != 1.

legendre_Ql

This function computes the Legendre function Q_l(x) for x > -1, x != 1 and l >= 0.

legendre_Ql_e

This function computes the Legendre function Q_l(x) for x > -1, x != 1 and l >= 0.