Module rgsl::physical_constant

Physical Constants

This chapter describes macros for the values of physical constants, such as the speed of light, c, and gravitational constant, G. The values are available in different unit systems, including the standard MKSA system (meters, kilograms, seconds, amperes) and the CGSM system (centimeters, grams, seconds, gauss), which is commonly used in Astronomy.

##References and Further Reading

The authoritative sources for physical constants are the 2006 CODATA recommended values, published in the article below. Further information on the values of physical constants is also available from the NIST website.

P.J. Mohr, B.N. Taylor, D.B. Newell, “CODATA Recommended Values of the Fundamental Physical Constants: 2006”, Reviews of Modern Physics, 80(2), pp. 633–730 (2008).

http://www.physics.nist.gov/cuu/Constants/index.html http://physics.nist.gov/Pubs/SP811/appenB9.html

Statics

MKSA_ACRE
The area of 1 acre. m^2

MKSA_ANGSTROM
The length of 1 angstrom. m

MKSA_ASTRONOMICAL_UNIT
The length of 1 astronomical unit (mean earth-sun distance), au. m

MKSA_BAR
The pressure of 1 bar. kg / m s^2

MKSA_BARN
The area of 1 barn. m^2

MKSA_BOHR_MAGNETON
The Bohr Magneton, \mu_B. A m^2

MKSA_BOHR_RADIUS
The Bohr radius, a_0. m

MKSA_BOLTZMANN
The Boltzmann constant, k. kg m^2 / K s^2

MKSA_BTU
The energy of 1 British Thermal Unit, btu. kg m^2 / s^2

MKSA_CALORIE
The energy of 1 calorie. kg m^2 / s^2

MKSA_CANADIAN_GALLON
The volume of 1 Canadian gallon. m^3

MKSA_CARAT
The mass of 1 carat. kg

MKSA_CUP
m^3

MKSA_CURIE
The activity of 1 curie. 1 / s

MKSA_DAY
The number of seconds in 1 day. s

MKSA_DEBYE
The electric dipole moment of 1 Debye, D. A s^2 / m^2

MKSA_DYNE
The force of 1 Dyne = 10^-5 Newton. kg m / s^2

MKSA_ELECTRON_CHARGE
The charge of the electron, e. A s

MKSA_ELECTRON_MAGNETIC_MOMENT
The absolute value of the magnetic moment of the electron, \mu_e. The physical magnetic moment of the electron is negative. A m^2

MKSA_ELECTRON_VOLT
The energy of 1 electron volt, eV. kg m^2 / s^2

MKSA_ERG
The energy 1 erg = 10^-7 Joule. kg m^2 / s^2

MKSA_FARADAY
The molar charge of 1 Faraday. A s / mol

MKSA_FATHOM
The length of 1 fathom. m

MKSA_FOOT
The length of 1 foot. m

MKSA_FOOTCANDLE
The illuminance of 1 footcandle. cd sr / m^2

MKSA_FOOTLAMBERT
The luminance of 1 footlambert. cd sr / m^2

MKSA_GAUSS
The magnetic field of 1 Gauss. kg / A s^2

MKSA_GRAM_FORCE
The force of 1 gram weight. kg m / s^2

MKSA_GRAVITATIONAL_CONSTANT
The gravitational constant, G. m^3 / kg s^2

MKSA_GRAV_ACCEL
The standard gravitational acceleration on Earth, g. m / s^2

MKSA_HECTARE
The area of 1 hectare. m^2

MKSA_HORSEPOWER
The power of 1 horsepower. kg m^2 / s^3

MKSA_HOUR
The number of seconds in 1 hour. s

MKSA_INCH
The length of 1 inch. m

MKSA_INCH_OF_MERCURY
The pressure of 1 inch of mercury. kg / m s^2

MKSA_INCH_OF_WATER
The pressure of 1 inch of water. kg / m s^2

MKSA_JOULE
The SI unit of energy, 1 Joule. kg m^2 / s^2

MKSA_KILOMETERS_PER_HOUR
The speed of 1 kilometer per hour. m / s

MKSA_KILOPOUND_FORCE
The force of 1 kilopound weight. kg m / s^2

MKSA_KNOT
The speed of 1 knot. m / s

MKSA_LAMBERT
The luminance of 1 lambert. cd sr / m^2

MKSA_LIGHT_YEAR
The distance of 1 light-year, ly. m

MKSA_LITER
The volume of 1 liter. m^3

MKSA_LUMEN
The luminous flux of 1 lumen. cd sr

MKSA_LUX
The illuminance of 1 lux. cd sr / m^2

MKSA_MASS_ELECTRON
The mass of the electron, m_e. kg

MKSA_MASS_MUON
The mass of the muon, m_\mu. kg

MKSA_MASS_NEUTRON
The mass of the neutron, m_n. kg

MKSA_MASS_PROTON
The mass of the proton, m_p. kg

MKSA_METER_OF_MERCURY
The pressure of 1 meter of mercury. kg / m s^2

MKSA_METRIC_TON
The mass of 1 metric ton (1000 kg). kg

MKSA_MICRON
The length of 1 micron. m

MKSA_MIL
The length of 1 mil (1/1000th of an inch). m

MKSA_MILE
The length of 1 mile. m

MKSA_MILES_PER_HOUR
The speed of 1 mile per hour. m / s

MKSA_MINUTE
The number of seconds in 1 minute. s

MKSA_MOLAR_GAS
The molar gas constant, R_0. kg m^2 / K mol s^2

MKSA_NAUTICAL_MILE
The length of 1 nautical mile. m

MKSA_NEWTON
The SI unit of force, 1 Newton. kg m / s^2

MKSA_NUCLEAR_MAGNETON
The Nuclear Magneton, \mu_N. A m^2

MKSA_OUNCE_MASS
The mass of 1 ounce. kg

MKSA_PARSEC
The distance of 1 parsec, pc. m

MKSA_PHOT
The illuminance of 1 phot. cd sr / m^2

MKSA_PINT
The volume of 1 pint. m^3

MKSA_PLANCKS_CONSTANT_H
Planck’s constant, h. kg m^2 / s

MKSA_PLANCKS_CONSTANT_HBAR
Planck’s constant divided by 2\pi, \hbar. kg m^2 / s

MKSA_POINT
The length of 1 printer’s point (1/72 inch). m

MKSA_POISE
The dynamic viscosity of 1 poise. kg m^-1 s^-1

MKSA_POUNDAL
The force of 1 poundal. kg m / s^2

MKSA_POUND_FORCE
The force of 1 pound weight. kg m / s^2

MKSA_POUND_MASS
The mass of 1 pound. kg

MKSA_PROTON_MAGNETIC_MOMENT
The magnetic moment of the proton, \mu_p. A m^2

MKSA_PSI
The pressure of 1 pound per square inch. kg / m s^2

MKSA_QUART
The volume of 1 quart. m^3

MKSA_RAD
The absorbed dose of 1 rad. m^2 / s^2

MKSA_ROENTGEN
The exposure of 1 roentgen. A s / kg

MKSA_RYDBERG
The Rydberg constant, Ry, in units of energy. This is related to the Rydberg inverse wavelength R_\infty by Ry = h c R_\infty. kg m^2 / s^2

MKSA_SOLAR_MASS
The mass of the Sun. kg

MKSA_SPEED_OF_LIGHT
The speed of light in vacuum, c. m / s

MKSA_STANDARD_GAS_VOLUME
The standard gas volume, V_0. m^3 / mol

MKSA_STD_ATMOSPHERE
The pressure of 1 standard atmosphere. kg / m s^2

MKSA_STEFAN_BOLTZMANN_CONSTANT
The Stefan-Boltzmann radiation constant, \sigma. kg / K^4 s^3

MKSA_STILB
The luminance of 1 stilb. cd / m^2

MKSA_STOKES
The kinematic viscosity of 1 stokes. m^2 / s

MKSA_TEXPOINT
The length of 1 TeX point (1/72.27 inch). m

MKSA_THERM
The energy of 1 Therm. kg m^2 / s^2

MKSA_THOMSON_CROSS_SECTION
The Thomson cross section, \sigma_T. m^2

MKSA_TON
The mass of 1 ton. kg

MKSA_TORR
The pressure of 1 torr. kg / m s^2

MKSA_TROY_OUNCE
The mass of 1 troy ounce. kg

MKSA_UK_GALLON
The volume of 1 UK gallon. m^3

MKSA_UK_TON
The mass of 1 UK ton. kg

MKSA_UNIFIED_ATOMIC_MASS
The unified atomic mass, amu. kg

MKSA_US_GALLON
The volume of 1 US gallon. m^3

MKSA_VACUUM_PERMEABILITY
The permeability of free space, \mu_0. This constant is defined in the MKSA system only. kg m / A^2 s^2

MKSA_VACUUM_PERMITTIVITY
The permittivity of free space, \epsilon_0. This constant is defined in the MKSA system only. A^2 s^4 / kg m^3

MKSA_WEEK
The number of seconds in 1 week. s

MKSA_YARD
The length of 1 yard. m

NUM_ATTO
10^-18

NUM_AVOGADRO
Avogadro’s number, N_a. 1 / mol

NUM_EXA
10^18

NUM_FEMTO
10^-15

NUM_FINE_STRUCTURE
The electromagnetic fine structure constant \alpha. 1

NUM_GIGA
10^9

NUM_KILO
10^3

NUM_MEGA
10^6

NUM_MICRO
10^-6

NUM_MILLI
10^-3

NUM_NANO
10^-9

NUM_PETA
10^15

NUM_PICO
10^-12

NUM_TERA
10^12

NUM_YOCTO
10^-24

NUM_YOTTA
10^24

NUM_ZEPTO
10^-21

NUM_ZETTA
10^21