Module rgsl::polynomials::evaluation

The functions described here evaluate the polynomial P(x) = c[0] + c[1] x + c[2] x^2 + \dots + c[len-1] x^{len-1} using Horner’s method for stability.

Functions

complex_poly_complex_eval
This function evaluates a polynomial with complex coefficients for the complex variable z.

poly_complex_eval
This function evaluates a polynomial with real coefficients for the complex variable z.

poly_eval
This function evaluates a polynomial with real coefficients for the real variable x.

poly_eval_derivs
This function evaluates a polynomial and its derivatives storing the results in the array res of size lenres. The output array contains the values of d^k P/d x^k for the specified value of x starting with k = 0.