Trait rgsl::pow::Pow

pub trait Pow {
    fn pow_int(&self, n: i32) -> Self;

    fn pow_uint(&self, n: u32) -> Self;

    fn pow2(&self) -> Self;

    fn pow3(&self) -> Self;

    fn pow4(&self) -> Self;

    fn pow5(&self) -> Self;

    fn pow6(&self) -> Self;

    fn pow7(&self) -> Self;

    fn pow8(&self) -> Self;

    fn pow9(&self) -> Self;
}

Required methods

fn pow_int(&self, n: i32) -> Self

This routine computes the power x^n for integer n. The power is computed efficiently—for example, x^8 is computed as ((x^2)^2)^2, requiring only 3 multiplications. A version of this function which also computes the numerical error in the result is available as gsl_sf_pow_int_e.

fn pow_uint(&self, n: u32) -> Self

This routine computes the power x^n for integer n. The power is computed efficiently—for example, x^8 is computed as ((x^2)^2)^2, requiring only 3 multiplications. A version of this function which also computes the numerical error in the result is available as gsl_sf_pow_int_e.

fn pow2(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

fn pow3(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

fn pow4(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

fn pow5(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

fn pow6(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

fn pow7(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

fn pow8(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

fn pow9(&self) -> Self

This function can be used to compute small integer powers x^2, x^3, etc. efficiently.

Loading content...