Module rgsl::randist::beta

Functions

beta
This function returns a random variate from the beta distribution. The distribution function is,

beta_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the beta distribution with parameters a and b.

beta_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the beta distribution with parameters a and b.

beta_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the beta distribution with parameters a and b.

beta_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the beta distribution with parameters a and b.

beta_pdf
This function computes the probability density p(x) at x for a beta distribution with parameters a and b, using the formula given above.