Module rgsl::randist::chi_squared

The chi-squared distribution arises in statistics. If Y_i are n independent Gaussian random variates with unit variance then the sum-of-squares,

X_i = \sum_i Y_i^2

has a chi-squared distribution with n degrees of freedom.

Functions

chisq
This function returns a random variate from the chi-squared distribution with nu degrees of freedom. The distribution function is,

chisq_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the chi-squared distribution with nu degrees of freedom.

chisq_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the chi-squared distribution with nu degrees of freedom.

chisq_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the chi-squared distribution with nu degrees of freedom.

chisq_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the chi-squared distribution with nu degrees of freedom.

chisq_pdf
This function computes the probability density p(x) at x for a chi-squared distribution with nu degrees of freedom, using the formula given above.