Module rgsl::randist::gumbel::type_2

Functions

gumbel2
This function returns a random variate from the Type-2 Gumbel distribution. The Type-2 Gumbel distribution function is,

gumbel2_pdf
This function computes the probability density p(x) at x for a Type-2 Gumbel distribution with parameters a and b, using the formula given above.

gumbel2_P
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Type-2 Gumbel distribution with parameters a and b.

gumbel2_Q
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Type-2 Gumbel distribution with parameters a and b.

gumbel2_Pinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Type-2 Gumbel distribution with parameters a and b.

gumbel2_Qinv
This function computes the cumulative distribution functions P(x), Q(x) and their inverses for the Type-2 Gumbel distribution with parameters a and b.