Function rgsl::statistics::quantile_from_sorted_data

pub fn quantile_from_sorted_data(
    data: &[f64], 
    stride: usize, 
    n: usize, 
    f: f64
) -> f64

This function returns a quantile value of sorted_data, a double-precision array of length n with stride stride. The elements of the array must be in ascending numerical order. The quantile is determined by the f, a fraction between 0 and 1. For example, to compute the value of the 75th percentile f should have the value 0.75.

There are no checks to see whether the data are sorted, so the function gsl_sort should always be used first.

The quantile is found by interpolation, using the formula

quantile = (1 - \delta) x_i + \delta x_{i+1}

where i is floor((n - 1)f) and \delta is (n-1)f - i.

Thus the minimum value of the array (data[0*stride]) is given by f equal to zero, the maximum value (data[(n-1)*stride]) is given by f equal to one and the median value is given by f equal to 0.5. Since the algorithm for computing quantiles involves interpolation this function always returns a floating-point number, even for integer data types.