Module rgsl::types

Re-exports

pub use self::basis_spline::BSpLineWorkspace;

pub use self::chebyshev::ChebSeries;

pub use self::combination::Combination;

pub use self::complex::ComplexF32;

pub use self::complex::ComplexF64;

pub use self::discrete_hankel::DiscreteHankel;

pub use self::eigen_symmetric_workspace::EigenGenHermVWorkspace;

pub use self::eigen_symmetric_workspace::EigenGenHermWorkspace;

pub use self::eigen_symmetric_workspace::EigenGenSymmVWorkspace;

pub use self::eigen_symmetric_workspace::EigenGenSymmWorkspace;

pub use self::eigen_symmetric_workspace::EigenGenVWorkspace;

pub use self::eigen_symmetric_workspace::EigenGenWorkspace;

pub use self::eigen_symmetric_workspace::EigenHermitianVWorkspace;

pub use self::eigen_symmetric_workspace::EigenHermitianWorkspace;

pub use self::eigen_symmetric_workspace::EigenNonSymmetricVWorkspace;

pub use self::eigen_symmetric_workspace::EigenNonSymmetricWorkspace;

pub use self::eigen_symmetric_workspace::EigenSymmetricVWorkspace;

pub use self::eigen_symmetric_workspace::EigenSymmetricWorkspace;

pub use self::fast_fourier_transforms::FftComplexF32WaveTable;

pub use self::fast_fourier_transforms::FftComplexF32Workspace;

pub use self::fast_fourier_transforms::FftComplexF64WaveTable;

pub use self::fast_fourier_transforms::FftComplexF64Workspace;

pub use self::histograms::Histogram;

pub use self::histograms::Histogram2D;

pub use self::histograms::Histogram2DPdf;

pub use self::histograms::HistogramPdf;

pub use self::integration::CquadWorkspace;

pub use self::integration::GLFixedTable;

pub use self::integration::IntegrationFixedType;

pub use self::integration::IntegrationFixedWorkspace;

pub use self::integration::IntegrationQawoTable;

pub use self::integration::IntegrationQawsTable;

pub use self::integration::IntegrationWorkspace;

pub use self::interpolation::Interp;

pub use self::interpolation::InterpAccel;

pub use self::interpolation::InterpType;

pub use self::interpolation::Spline;

pub use self::mathieu::MathieuWorkspace;

pub use self::matrix::MatrixF32;

pub use self::matrix::MatrixF32View;

pub use self::matrix::MatrixF64;

pub use self::matrix::MatrixF64View;

pub use self::matrix::MatrixI32;

pub use self::matrix::MatrixI32View;

pub use self::matrix::MatrixU32;

pub use self::matrix::MatrixU32View;

pub use self::matrix_complex::MatrixComplexF32;

pub use self::matrix_complex::MatrixComplexF64;

pub use self::minimizer::Minimizer;

pub use self::minimizer::MinimizerType;

pub use self::monte_carlo::MiserMonteCarlo;

pub use self::monte_carlo::MiserParams;

pub use self::monte_carlo::PlainMonteCarlo;

pub use self::monte_carlo::VegasMonteCarlo;

pub use self::monte_carlo::VegasParams;

pub use self::multifit_linear::MultifitLinearWorkspace;

pub use self::multifit_solver::MultiFitFdfSolver;

pub use self::multifit_solver::MultiFitFdfSolverType;

pub use self::multifit_solver::MultiFitFunction;

pub use self::multifit_solver::MultiFitFunctionFdf;

pub use self::multiset::MultiSet;

pub use self::n_tuples::ReadNTuples;

pub use self::n_tuples::WriteNTuples;

pub use self::ordinary_differential_equations::ODEiv2Control;

pub use self::ordinary_differential_equations::ODEiv2Driver;

pub use self::ordinary_differential_equations::ODEiv2Evolve;

pub use self::ordinary_differential_equations::ODEiv2Step;

pub use self::ordinary_differential_equations::ODEiv2StepType;

pub use self::ordinary_differential_equations::ODEiv2System;

pub use self::permutation::Permutation;

pub use self::polynomial::PolyComplexWorkspace;

pub use self::qrng::QRng;

pub use self::qrng::QRngType;

pub use self::ran_discrete::RanDiscrete;

pub use self::result::Result;

pub use self::result::ResultE10;

pub use self::rng::Rng;

pub use self::rng::RngType;

pub use self::roots::RootFSolver;

pub use self::roots::RootFSolverType;

pub use self::roots::RootFdfSolver;

pub use self::roots::RootFdfSolverType;

pub use self::rstat::RStatQuantileWorkspace;

pub use self::rstat::RStatWorkspace;

pub use self::series_acceleration::LevinUTruncWorkspace;

pub use self::series_acceleration::LevinUWorkspace;

pub use self::siman::SimAnnealing;

pub use self::siman::SimAnnealingParams;

pub use self::vector::VectorF32;

pub use self::vector::VectorF32View;

pub use self::vector::VectorF64;

pub use self::vector::VectorF64View;

pub use self::vector::VectorI32;

pub use self::vector::VectorI32View;

pub use self::vector::VectorU32;

pub use self::vector::VectorU32View;

pub use self::vector_complex::VectorComplexF32;

pub use self::vector_complex::VectorComplexF32View;

pub use self::vector_complex::VectorComplexF64;

pub use self::vector_complex::VectorComplexF64View;

pub use self::wavelet_transforms::Wavelet;

pub use self::wavelet_transforms::WaveletType;

pub use self::wavelet_transforms::WaveletWorkspace;

Modules

basis_spline

B-splines are commonly used as basis functions to fit smoothing curves to large data sets. To do this, the abscissa axis is broken up into some number of intervals, where the endpoints of each interval are called breakpoints.

chebyshev

Chebyshev Approximations

combination

Combinations

complex

discrete_hankel

Discrete Hankel Transforms

eigen_symmetric_workspace

Real Symmetric Matrices

fast_fourier_transforms

histograms

Histograms

integration

interpolation

Interpolation

mathieu

The routines described in this section compute the angular and radial Mathieu functions, and their characteristic values. Mathieu functions are the solutions of the following two differential equations:

matrix

Matrices

matrix_complex

minimizer

One dimensional Minimization

monte_carlo

Monte Carlo Integration

multifit_linear

multifit_solver

Nonlinear Least-Squares Fitting

multiroot

Multi dimensional Root-Finding

multiset

Multisets

n_tuples

N-tuples

ordinary_differential_equations

Numerical ODE solvers.

permutation

polynomial

General Polynomial Equations

qrng

Quasi-Random Sequences

ran_discrete

Given K discrete events with different probabilities P[k], produce a random value k consistent with its probability.

result

rng

Random Number Generation

roots

One dimensional Root-Finding

rstat

series_acceleration

Series Acceleration

siman

25 Simulated Annealing

vector

Vectors

vector_complex

wavelet_transforms

Wavelet Transforms