Struct rgsl::types::matrix::MatrixView

pub struct MatrixView { /* fields omitted */ }

Methods

impl MatrixView

pub fn from_matrix(
m: &mut MatrixF64,
k1: usize,
k2: usize,
n1: usize,
n2: usize
) -> MatrixView

These functions return a matrix view of a submatrix of the matrix m. The upper-left element of the submatrix is the element (k1,k2) of the original matrix. The submatrix has n1 rows and n2 columns. The physical number of columns in memory given by tda is unchanged. Mathematically, the (i,j)-th element of the new matrix is given by,

m'(i,j) = m->data[(k1m->tda + k2) + im->tda + j]

where the index i runs from 0 to n1-1 and the index j runs from 0 to n2-1.

The data pointer of the returned matrix struct is set to null if the combined parameters (i,j,n1,n2,tda) overrun the ends of the original matrix.

The new matrix view is only a view of the block underlying the existing matrix, m. The block containing the elements of m is not owned by the new matrix view. When the view goes out of scope the original matrix m and its block will continue to exist. The original memory can only be deallocated by freeing the original matrix. Of course, the original matrix should not be deallocated while the view is still in use.

The function gsl_matrix_const_submatrix is equivalent to gsl_matrix_submatrix but can be used for matrices which are declared const.

pub fn from_array(base: &mut [f64], n1: usize, n2: usize) -> MatrixView

These functions return a matrix view of the array base. The matrix has n1 rows and n2 columns. The physical number of columns in memory is also given by n2. Mathematically, the (i,j)-th element of the new matrix is given by,

m'(i,j) = base[i*n2 + j]

where the index i runs from 0 to n1-1 and the index j runs from 0 to n2-1.

The new matrix is only a view of the array base. When the view goes out of scope the original array base will continue to exist. The original memory can only be deallocated by freeing the original array. Of course, the original array should not be deallocated while the view is still in use.

The function gsl_matrix_const_view_array is equivalent to gsl_matrix_view_array but can be used for matrices which are declared const.

pub fn from_array_with_tda(
base: &mut [f64],
n1: usize,
n2: usize,
tda: usize
) -> MatrixView

These functions return a matrix view of the array base with a physical number of columns tda which may differ from the corresponding dimension of the matrix. The matrix has n1 rows and n2 columns, and the physical number of columns in memory is given by tda. Mathematically, the (i,j)-th element of the new matrix is given by,

m'(i,j) = base[i*tda + j]

where the index i runs from 0 to n1-1 and the index j runs from 0 to n2-1.

The new matrix is only a view of the array base. When the view goes out of scope the original array base will continue to exist. The original memory can only be deallocated by freeing the original array. Of course, the original array should not be deallocated while the view is still in use.

The function gsl_matrix_const_view_array_with_tda is equivalent to gsl_matrix_view_array_with_tda but can be used for matrices which are declared const.

pub fn from_vector(v: &mut VectorF64, n1: usize, n2: usize) -> MatrixView

These functions return a matrix view of the vector v. The matrix has n1 rows and n2 columns. The vector must have unit stride. The physical number of columns in memory is also given by n2. Mathematically, the (i,j)-th element of the new matrix is given by,

m'(i,j) = v->data[i*n2 + j]

where the index i runs from 0 to n1-1 and the index j runs from 0 to n2-1.

The new matrix is only a view of the vector v. When the view goes out of scope the original vector v will continue to exist. The original memory can only be deallocated by freeing the original vector. Of course, the original vector should not be deallocated while the view is still in use.

The function gsl_matrix_const_view_vector is equivalent to gsl_matrix_view_vector but can be used for matrices which are declared const.

pub fn from_vector_with_tda(
v: &mut VectorF64,
n1: usize,
n2: usize,
tda: usize
) -> MatrixView

These functions return a matrix view of the vector v with a physical number of columns tda which may differ from the corresponding matrix dimension. The vector must have unit stride. The matrix has n1 rows and n2 columns, and the physical number of columns in memory is given by tda. Mathematically, the (i,j)-th element of the new matrix is given by,

m'(i,j) = v->data[i*tda + j]

where the index i runs from 0 to n1-1 and the index j runs from 0 to n2-1.

The new matrix is only a view of the vector v. When the view goes out of scope the original vector v will continue to exist. The original memory can only be deallocated by freeing the original vector. Of course, the original vector should not be deallocated while the view is still in use.

The function gsl_matrix_const_view_vector_with_tda is equivalent to gsl_matrix_view_vector_with_tda but can be used for matrices which are declared const.

pub fn matrix(&mut self) -> MatrixF64