Struct rgsl::types::multifit_solver::MultiFitFdfSolver

pub struct MultiFitFdfSolver { /* fields omitted */ }

Methods

impl MultiFitFdfSolver

pub fn new(
_type: &MultiFitFdfSolverType,
n: usize,
p: usize
) -> Option<MultiFitFdfSolver>

This function returns a pointer to a newly allocated instance of a solver of type T for n observations and p parameters. The number of observations n must be greater than or equal to parameters p.

pub fn set(&mut self, f: &mut MultiFitFunctionFdf, x: &VectorF64) -> Value

This function initializes, or reinitializes, an existing solver s to use the function f and the initial guess x.

pub fn x(&self) -> VectorF64

pub fn f(&self) -> VectorF64

pub fn J(&self) -> MatrixF64

pub fn dx(&self) -> VectorF64

pub fn name(&self) -> String

pub fn iterate(&mut self) -> Value

This function performs a single iteration of the solver s. If the iteration encounters an unexpected problem then an error code will be returned. The solver maintains a current estimate of the best-fit parameters at all times.

pub fn position(&self) -> VectorF64

This function returns the current position (i.e. best-fit parameters) s->x of the solver s.

pub fn driver(&mut self, max_iter: usize, epsabs: f64, epsrel: f64) -> Value

These functions iterate the solver s for a maximum of maxiter iterations. After each iteration, the system is tested for convergence using gsl_multifit_test_delta with the error tolerances epsabs and epsrel.