Module rgsl::types::rng::unix

The standard Unix random number generators rand, random and rand48 are provided as part of GSL. Although these generators are widely available individually often they aren’t all available on the same platform. This makes it difficult to write portable code using them and so we have included the complete set of Unix generators in GSL for convenience. Note that these generators don’t produce high-quality randomness and aren’t suitable for work requiring accurate statistics. However, if you won’t be measuring statistical quantities and just want to introduce some variation into your program then these generators are quite acceptable.

Functions

rand

This is the BSD rand generator. Its sequence is

rand48

This is the Unix rand48 generator. Its sequence is

random_bsd

These generators implement the random family of functions, a set of linear feedback shift register generators originally used in BSD Unix. There are several versions of random in use today: the original BSD version (e.g. on SunOS4), a libc5 version (found on older GNU/Linux systems) and a glibc2 version. Each version uses a different seeding procedure, and thus produces different sequences.

random_glibc2

These generators implement the random family of functions, a set of linear feedback shift register generators originally used in BSD Unix. There are several versions of random in use today: the original BSD version (e.g. on SunOS4), a libc5 version (found on older GNU/Linux systems) and a glibc2 version. Each version uses a different seeding procedure, and thus produces different sequences.

random_libc5

These generators implement the random family of functions, a set of linear feedback shift register generators originally used in BSD Unix. There are several versions of random in use today: the original BSD version (e.g. on SunOS4), a libc5 version (found on older GNU/Linux systems) and a glibc2 version. Each version uses a different seeding procedure, and thus produces different sequences.