Module rgsl::zeta

The Riemann zeta function is defined in Abramowitz & Stegun, Section 23.2.

Modules

eta
The eta function is defined by \eta(s) = (1-2^{1-s}) \zeta(s).

hurwitz
The Hurwitz zeta function is defined by \zeta(s,q) = \sum_0^\infty (k+q)^{-s}.

riemann
The Riemann zeta function is defined by the infinite sum \zeta(s) = \sum_{k=1}^\infty k^{-s}.

riemann_mins_one
For large positive argument, the Riemann zeta function approaches one. In this region the fractional part is interesting, and therefore we need a function to evaluate it explicitly.